% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R,
%   R/class_spflow-model-meta.R
\name{add_details}
\alias{add_details}
\title{Internal method to add details to a \code{\link[=spflow_model-class]{spflow_model-class()}}}
\usage{
add_details(object, ...)

\S4method{add_details}{spflow_model}(object, model_matrices, flow_control, model_moments)
}
\arguments{
\item{object}{A \code{\link[=spflow_model-class]{spflow_model-class()}}}

\item{model_matrices}{A list as returned by \code{\link[=spflow_model_matrix]{spflow_model_matrix()}}}

\item{flow_control}{A list as returned by \code{\link[=spflow_control]{spflow_control()}}}

\item{model_moments}{A list as returned by \code{\link[=spflow_model_moments]{spflow_model_moments()}}}
}
\value{
An object of class spflow_model
}
\description{
Internal method to add details to a \code{\link[=spflow_model-class]{spflow_model-class()}}
}
\details{
The method adds the design matrix and the coefficient names to an
\code{\link[=spflow_model-class]{spflow_model-class()}} object.
It also calculates the fitted values and the residuals as well as a
goodness-of-fit measure.
}
\keyword{internal}
