% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_error.R
\name{assert_errors}
\alias{assert_errors}
\alias{assert}
\alias{assert_is}
\alias{assert_inherits}
\alias{assert_is_one_of}
\alias{assert_is_single_x}
\alias{assert_valid_case}
\title{A set of functions to create customized error messages}
\usage{
assert(expr, error_msg = "ERROR", ..., warn = FALSE)

assert_is(obj, .class)

assert_inherits(obj, .class)

assert_is_one_of(obj, .classes)

assert_is_single_x(obj, x)

assert_valid_case(argument, cases)
}
\value{
Functions called for their side effect, raises warning or error
}
\description{
The functions are build on top of \code{\link[=sprintf]{sprintf()}}
}
\author{
Lukas Dargel
}
\keyword{internal}
