% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-model-matrix_2-transformations.R
\name{by_source_model_matrix}
\alias{by_source_model_matrix}
\title{Internal functions to generate model matrices}
\usage{
by_source_model_matrix(formula_parts, data_sources)
}
\value{
A list of matrices
}
\description{
Internal functions to generate model matrices
}
\details{
Sources describe the data.frames holding the original information on
the nodes and the node pairs. There are three possible source data.frame
which are referred to as "pair", "orig", or "dest".
All formulas (normal, sdm, and instrument) are combined to generate an
overall model matrix which is only expanded once for each source.
}
\keyword{internal}
