% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_model.R
\name{compare_results}
\alias{compare_results}
\title{Compare results of multiple spflow_models}
\usage{
compare_results(
  model_list,
  global_vars = c("model_coherence", "R2_corr", "ll", "AIC", "N_sample")
)
}
\arguments{
\item{model_list}{a list of models that should be compared (only \code{\link[=spflow_model-class]{spflow_model-class()}} is used)}

\item{global_vars}{a character indicating which statistics should be reported}

\item{digits}{a numeric indicating to what decimal the results should be rounded}

\item{sig_levels}{a named numeric indicating the codification of significance levels}

\item{add_dispersion}{a logical, if \code{TRUE} the standard errors are added in parenthesis}
}
\value{
a data.frame
}
\description{
Compare results of multiple spflow_models
}
\examples{

res_ge  <- spflow(y9 ~ . + P_(DISTANCE), multi_net_usa_ge, "ge_ge")
res_usa <- spflow(y9 ~ . + P_(DISTANCE), multi_net_usa_ge, "usa_usa")
compare_results(list("GE" = res_ge, "US" = res_usa),
                global_vars = c("N_sample", "R2_corr", "model_coherence")) )
}
\author{
Lukas Dargel
}
