% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics_and_maybes.R
\name{complete_pairs}
\alias{complete_pairs}
\title{Generic for completing the missing od-pairs in a \code{\link[=spflow_multinet]{spflow_multinet()}}}
\usage{
complete_pairs(object, ...)
}
\arguments{
\item{object}{A \code{\link[=spflow_model-class]{spflow_model-class()}}}

\item{...}{Arguments passed on to methods}
}
\value{
A \code{\link[=spflow_model-class]{spflow_model-class()}}
}
\description{
For details see the documentation of the corresponding class.
}
\seealso{
\code{\link[=spflow_model-class]{spflow_model-class()}}
}
