% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow_moments_1.R
\name{compute_spflow_moments}
\alias{compute_spflow_moments}
\title{Compute the principal moments for a spatial flow model}
\usage{
compute_spflow_moments(spflow_matrices, n_o, n_d, N, wt, na_rm = FALSE)
}
\description{
Compute the principal moments for a spatial flow model
}
\keyword{internal}
