% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_image.R
\name{cor_image}
\alias{cor_image}
\title{Graphical display of a correlation matrix}
\usage{
cor_image(cor_mat, cex = 0.7, srt = 45, cex.axis = 0.7)
}
\arguments{
\item{cor_mat}{A matrix of pairwise correlations}

\item{cex, cex.axis}{Numerics indicating the font size for text- and axis labels}

\item{srt}{A numeric controlling the angle of the text}
}
\value{
A graphical representation of the correlation matrix
}
\description{
The function shows an \code{image()} of a correlation matrix, where row-and column names
are use as labels.
}
\examples{
cor_mat <- cor(matrix(rnorm(40),ncol = 4))
cor_image(cor_mat)
}
\author{
Lukas Dargel
}
