% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\name{corr_map}
\alias{corr_map}
\title{Graphic display for correlation matrices}
\usage{
corr_map(cor_mat)
}
\arguments{
\item{cor_mat}{A matrix of pairwise correlations}
}
\value{
A graphical representation of the correlation matrix
}
\description{
Graphic display for correlation matrices
}
\examples{

cor_mat <- cor(matrix(rnorm(40),ncol = 4))
corr_map(cor_mat)
}
