% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-error.R
\name{custom_primitives}
\alias{custom_primitives}
\alias{is_one_of}
\alias{is_single_character}
\alias{is_single_logical}
\alias{has_equal_elements}
\alias{has_distinct_elements}
\title{A set of primitive functions to test conditions}
\usage{
is_one_of(.obj, .classes)

is_single_character(x)

is_single_logical(x)

has_equal_elements(obj)

has_distinct_elements(obj)
}
\description{
A set of primitive functions to test conditions
}
\author{
Lukas Dargel
}
\keyword{internal}
