% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{dat}
\alias{dat}
\alias{dat-set}
\alias{dat<-}
\title{Access the data of \link[=sp_network_classes]{sp network objects}}
\arguments{
\item{object}{One of \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}}, \code{\link[=sp_network_pair-class]{sp_network_pair-class()}}}

\item{value}{A data.frame to replace existing data}
}
\description{
For details see the documentation of the corresponding classes.
}
