% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R,
%   R/class_sp-network-meta.R
\name{dat_template}
\alias{dat_template}
\alias{dat_template,sp_network_nodes_pairs-method}
\title{Internal method that allows to extract the schema of the data inside
\code{\link[=sp_network_classes]{sp_network_classes()}}}
\usage{
\S4method{dat_template}{sp_network_nodes_pairs}(object, col_subset = NULL)
}
\arguments{
\item{object}{A \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}} or \code{\link[=sp_network_pair-class]{sp_network_pair-class()}}}

\item{col_subset}{A character vector indicating columns that should be kept for the template}
}
\description{
The method extract the columns of the data with zero rows which allows to
use it as a template for other functions or to extract the column names.
}
\keyword{internal}
