% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-model-matrix_1.R
\name{define_variable_roles}
\alias{define_variable_roles}
\title{Internal functions to generate model matrices}
\usage{
define_variable_roles(formula_parts, data_sources)
}
\value{
A list of lists, with information on the way variables are used
in the model. (e.g. origin vs destination and the number of lags)
}
\description{
Internal functions to generate model matrices
}
\details{
The function generates a list for each role \code{c("Y_","G_","O_","D_","I_")}
which indicates whether a variable in the design matrix is used as
\code{c("norm", "sdm", "inst")}.
}
\keyword{internal}
