% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow_matrices.R
\name{derive_spflow_matrices}
\alias{derive_spflow_matrices}
\title{Generate the design matrices for the spatial interaction model}
\usage{
derive_spflow_matrices(
  spflow_data,
  spflow_neighborhood,
  spflow_formula,
  spflow_control,
  na_rm = FALSE
)
}
\value{
A list of design matrices for the spatial interaction model
}
\description{
Create the model matrices required for matrix-form estimation of
the spatial econometric interaction model.
}
\details{
The key to an efficient estimation is to preserve the relational
representation of the data for origins, destinations and
origins-destinations pairs.
This requires to be aware of the;
\itemize{
\item three sources of data (pair, orig, dest)
\item three parts of the formula (norm, sdm, inst)
\item five roles of the variables (Y_, P_, D_, O_, I_)
}

The additional separation of data sources and roles makes sense if the list
of origins coincides with the list of destinations.
In this case, we can use data from the same source as origin, destination,
or intra-regional characteristics.

The model formulas contain information about mathematical transformations
of the variables and allow to deduce number of spatial lags required for
each of them.
To apply the transformations we use R's build-in tools for handling formulas.
Spatial lags are calculated after the transformations have been applied.
Below is an explanation of the formula parts:
\itemize{
\item norm variables are not lagged
\item sdm variables are lagged once and used as explanatory variables
\item inst variables are lagged twice and used as instruments.
If a variable is at the same time inst and sdm we have to increase the
lags-order to avoid duplicating columns
(see \insertCite{Dargel2021}{spflow}).
}
}
\references{
\insertAllCited{}
}
\keyword{internal}
