% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R,
%   R/class_sp-network-meta.R
\name{drop_columns}
\alias{drop_columns}
\alias{drop_columns,sp_network_nodes_pairs-method}
\title{Internal method that allows to drop columns from the data inside
\code{\link[=sp_network_classes]{sp_network_classes()}}}
\usage{
\S4method{drop_columns}{sp_network_nodes_pairs}(object, drop_columns)
}
\arguments{
\item{object}{A \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}} or \code{\link[=sp_network_pair-class]{sp_network_pair-class()}}}

\item{drop_columns}{A vector indicating columns of the data;
either character or integer.}
}
\description{
The function uses reference semantic and allows to manipulate the data
inside the object efficiently.
}
\keyword{internal}
