% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow_neighborhood.R
\name{drop_superfluent_nodes}
\alias{drop_superfluent_nodes}
\title{Remove origins or destinations from the neighborhood matrix that do
not appear in any of the OD pairs.}
\usage{
drop_superfluent_nodes(OW, DW, flow_indicator)
}
\description{
Remove origins or destinations from the neighborhood matrix that do
not appear in any of the OD pairs.
}
\keyword{internal}
