% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{elvis_operators}
\alias{elvis_operators}
\alias{\%||\%}
\alias{\%|!|\%}
\alias{\%T\%}
\title{Elvis operators}
\usage{
x \%||\% y

x \%|!|\% y

x \%T\% y
}
\arguments{
\item{x}{A value or Null}

\item{y}{alternative return}
}
\value{
x or y
}
\description{
Infix operators that reduce cyclomatic complexity of the source code.
In other words you will need fewer nested \code{if} statements.
}
\keyword{internal}
