% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow_control.R
\name{enhance_spflow_control}
\alias{enhance_spflow_control}
\title{Derive additional information from the \code{\link[=spflow_control]{spflow_control()}} arguments}
\usage{
enhance_spflow_control(estimation_control, is_within)
}
\value{
A list of control parameters
}
\description{
This is an internal function to validates the content of the
\code{estimation_control} argument of \code{\link[=spflow]{spflow()}}.
It also derives some verbal descriptions of the estimated model.
}
\author{
Lukas Dargel
}
\keyword{internal}
