% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\alias{multi_net_usa_ge}
\alias{germany_grid}
\alias{usa_net}
\alias{usa_grid}
\alias{germany_net}
\alias{simulation_params}
\title{Simulated data for stylized versions of Germany and the USA}
\format{
An object of class \code{spflow_network_multi} of length 1.

An object of class \code{sf} (inherits from \code{data.frame}) with 16 rows and 3 columns.

An object of class \code{spflow_network_nodes} of length 1.

An object of class \code{sf} (inherits from \code{data.frame}) with 51 rows and 3 columns.

An object of class \code{spflow_network_nodes} of length 1.

An object of class \code{list} of length 3.
}
\source{
The stylized versions of Germany and the USA are inspired by
https://ialab.it.monash.edu/~dwyer/papers/maptrix.pdf
}
\usage{
multi_net_usa_ge

germany_grid

usa_net

usa_grid

germany_net

simulation_params
}
\description{
The package provides two stylized country examples that have been already
used by\insertCite{Yang2017;textual}{spflow}.
The first example is a stylized version of Germany with 16 states and
second one is based on the USA with 51 states.

There are two \pkg{sf} objects that contain the geography and data for the country
examples.
The same information is also provided by two \code{\link[=spflow_network_nodes-class]{spflow_network_nodes-class()}} objects.
Additionally, there is an \code{\link[=spflow_network_multi-class]{spflow_network_multi-class()}} that contains the
two \code{\link[=spflow_network_nodes]{spflow_network_nodes()}} and four \code{\link[=spflow_network_pairs]{spflow_network_pairs()}}, which contain
OD-pairs where each states of each country can be origins or destinations.

For each of the OD-pairs we simulated three flow vectors \code{y1}, \code{y2} and \code{y9}.
The parameters used for this simulation are given as a list:
\code{rho} contains the auto-regression parameters, \code{delta} contains the parameters
for the exogenous variables and \code{sd_error} is the standard deviation of the Gaussian error term.
}
\references{
\insertAllCited{}
}
\author{
Lukas Dargel
}
\keyword{datasets}
