% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-flow-formula.R
\name{expand_case_formulas}
\alias{expand_case_formulas}
\title{Decomposes the formula provided to \code{\link[=spflow]{spflow()}} into four parts.}
\usage{
expand_case_formulas(rhs_formula)
}
\arguments{
\item{formula}{A formula which specifies the structural model}
}
\value{
A list of formulas for each data source.
}
\description{
The sub-formulas provided by this decomposition can be carried to the
diffrent data sources for exogenous data (origin data, destination data, pair data),
which permits to use a formular interface similar to the one provided by \code{\link[=lm]{lm()}}.
}
\keyword{internal}
