% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-flow-formula.R
\name{expand_roles_and_cases}
\alias{expand_roles_and_cases}
\title{Expand the structural formula of a spatial interaction model}
\usage{
expand_roles_and_cases(flow_forumula, flow_control, use_intra)
}
\arguments{
\item{flow_forumula}{A formulas corresponding to the structural interaction model}

\item{flow_control}{A \code{\link[=spflow_control]{spflow_control()}} list to fine tune the estimation}

\item{use_intra}{A logical which adds sperate set of coefficients for intra-observational flows (origin == destination) to the model.}
}
\value{

}
\description{
Decomposes the stractural formulas of a spatial interaction model into one
sided formuals which are then applied to five cases and in three diffrent
roles.
\itemize{
\item 5 cases: \strong{Interactions ~ Origins + Destinations + Intra + Pairs}
\item 3 roles: \strong{normal_variables sdm_variables instrumental_variables}
}
}
\details{
A spatial interaction model has potentially multiple data sources and
multiple roles for each variable.
The five cases refer to the type of data notably:
\enumerate{
\item response variable (interactions)
\item attributes of origins
\item attributes of destination
\item attributes of intra-observationl units
\item attributes or origin-destination pars (such as distance)
}

The three possible roles refer to the use of a variable as:
\enumerate{
\item normal variable
\item sdm variable (as spatial lag)
\item instrumental variable (to use during s2sls estimation)
}
}
