% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-logdet.R
\name{fodet1}
\alias{fodet1}
\title{Fast determinant calculation}
\usage{
fodet1(parms, traces, n, dev = T)
}
\arguments{
\item{parms}{?}

\item{traces}{?}

\item{n}{?}

\item{dev}{?}
}
\value{
?
}
\description{
Fast determinant calculation
}
\keyword{internal}
