% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow_logdet.R
\name{generate_approxldet_cartesian2}
\alias{generate_approxldet_cartesian2}
\title{Order 2 approximations for the cartesian case}
\usage{
generate_approxldet_cartesian2(OW = NULL, DW = NULL, n_o, n_d, model)
}
\description{
Order 2 approximations for the cartesian case
}
\keyword{internal}
