% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow_logdet.R
\name{generate_approxldet_noncartesian}
\alias{generate_approxldet_noncartesian}
\title{Order 4 approximations for non cartesian case}
\usage{
generate_approxldet_noncartesian(Wd, Wo, Ww, approx_order, model)
}
\description{
Order 4 approximations for non cartesian case
}
\keyword{internal}
