% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow_logdet.R
\name{generate_approxldet_noncartesian2}
\alias{generate_approxldet_noncartesian2}
\title{Order 2 approximations for the non-cartesian case}
\usage{
generate_approxldet_noncartesian2(
  OW = NULL,
  DW = NULL,
  M_indicator,
  n_o,
  n_d,
  model
)
}
\description{
Order 2 approximations for the non-cartesian case
}
\keyword{internal}
