% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{id}
\alias{id}
\alias{id-set}
\alias{id<-}
\title{Access the id of \link[=sp_network_classes]{spflow network objects}}
\arguments{
\item{object}{A \link[=sp_network_classes]{spflow network object}}

\item{value}{A character replacing the existing id}
}
\description{
For details see the documentation of the corresponding classes.
}
