% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fp.R
\name{\%p\%}
\alias{\%p\%}
\alias{\%T\%}
\alias{\%||\%}
\alias{\%|!|\%}
\title{Additional infix operators for more concise code}
\usage{
x \%p\% y

x \%T\% y

x \%||\% y

x \%|!|\% y
}
\description{
Additional infix operators for more concise code
}
\author{
Lukas Dargel
}
\keyword{internal}
