% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-simulations.R
\name{invert_spatial_filter}
\alias{invert_spatial_filter}
\title{Create an inverted spatial filter that can be used for simulations}
\usage{
invert_spatial_filter(weight_matrices, autoreg_parameters)
}
\arguments{
\item{weight_matrices}{A list of neighborhood matrices}

\item{autoreg_parameters}{A vector of parameters}
}
\value{
A matrix representing the inverted spatial filter
}
\description{
Create an inverted spatial filter that can be used for simulations
}
\seealso{
Other spflow simulation functions: 
\code{\link{expand_flow_neighborhood}()},
\code{\link{spatial_filter}()},
\code{\link{spflow_sim}()}
}
\concept{spflow simulation functions}
\keyword{internal}
