% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fp.R
\name{translist}
\alias{translist}
\alias{compact}
\alias{ilapply}
\alias{ulapply}
\alias{flatlist}
\title{Additional list operators that allow a more concise and
functional programming style}
\usage{
translist(.l)

compact(.x)

ilapply(.list, .f, MoreArgs = NULL)

ulapply(.list, .f, ..., recursive = TRUE, use.names = FALSE)

flatlist(lst, use.names = TRUE)
}
\description{
Additional list operators that allow a more concise and
functional programming style
}
\author{
Lukas Dargel
}
\keyword{internal}
