% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fp.R
\name{named_list}
\alias{named_list}
\alias{lookup}
\alias{list_lookup}
\alias{collect}
\alias{sort_names}
\title{Some helpers to create lookups that allow to program with key value pairs}
\usage{
named_list(names, init = NULL)

lookup(values, names = as.character(values))

list_lookup(values, names = as.character(values))

collect(names)

sort_names(x)
}
\description{
Some helpers to create lookups that allow to program with key value pairs
}
\author{
Lukas Dargel
}
\keyword{internal}
