% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics_and_maybes.R
\name{mcmc_results}
\alias{mcmc_results}
\title{Generic for accessing the MCMC results of a \code{\link[=spflow_model-class]{spflow_model-class()}}}
\usage{
mcmc_results(object)
}
\arguments{
\item{object}{A \code{\link[=spflow_model-class]{spflow_model-class()}}}
}
\value{
An mcmc object, containing the draws of the sampled parameters
}
\description{
For details see the documentation of the corresponding class.
}
\seealso{
\code{\link[=spflow_model-class]{spflow_model-class()}}
}
