% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-model-moments_2-blocks.R
\name{moment_empirical_var}
\alias{moment_empirical_var}
\title{Compute the moment expression in an interaction model in matrix form}
\usage{
moment_empirical_var(model_matrices, N, n_d, n_o)
}
\description{
Compute the moment expression in an interaction model in matrix form
}
\details{
The moment (U'U) moment matrix is grouped into (4x4) blocks.
These 16 blocks are derived as interactions from the four blocks
{alpha, alpha_I, beta, gamma}.
Only ten blocks are unique and the remaining six are inferred by symmetry.
The developments follow the theory presented in
\insertCite{Dargel2021;textual}{spflow}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
