% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{multi_net_usa_ge}
\alias{multi_net_usa_ge}
\alias{germany_grid}
\alias{usa_net}
\alias{usa_grid}
\alias{germany_net}
\title{Simulated data for stylized versions of Germany and the USA}
\format{
An object of class \code{sp_multi_network} of length 1.

An object of class \code{SpatialPolygonsDataFrame} with 16 rows and 2 columns.

An object of class \code{sp_network} of length 1.

An object of class \code{SpatialPolygonsDataFrame} with 51 rows and 2 columns.

An object of class \code{sp_network} of length 1.
}
\source{
inspired by \url{https://ialab.it.monash.edu/~dwyer/papers/maptrix.pdf.}
}
\usage{
multi_net_usa_ge

germany_grid

usa_net

usa_grid

germany_net
}
\description{
The first example is a stylized version of Germany with 16 states.
The second example is a stylized version of the USA with 51 states.
}
\details{
We provide the spatial objects which containt the geography of our example
as well as the prepared \code{\link[=sp_network]{sp_network()}} objects.
Finally there is a multinetwork object which contains the two networks and
four network pairs based on the same examples.
}
\keyword{datasets}
