% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\name{multinom_coef}
\alias{multinom_coef}
\title{Compute the multinomial coefficient}
\usage{
multinom_coef(...)
}
\description{
Compute the multinomial coefficient
}
\details{
The coefficient is computed for each row in a data.frame where the
rows correspond to the power and the columns for one element
}
\keyword{internal}
