% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_math.R
\name{multinom_table}
\alias{multinom_table}
\title{Create a table for  multinomial coefficient and parameter powers}
\usage{
multinom_table(max_power, coef_names)
}
\description{
Create a table for  multinomial coefficient and parameter powers
}
\keyword{internal}
