% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nb2Mat}
\alias{nb2Mat}
\title{Convert the \code{nb-class} from spdep to a sparse Matrix}
\usage{
nb2Mat(nb)
}
\arguments{
\item{nb}{A \code{nb-class} from spdep}
}
\value{
A sparse matrix
}
\description{
This function genereates a sparse Matrix from a neighborhood matrix.
It bypasses \code{spdep::nb2mat()} which leads to a dense matrix representation.
}
\examples{
nb <- spdep::poly2nb(germany_grid)
nb2Mat(nb)
}
\author{
Lukas Dargel
}
