% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{neighborhood}
\alias{neighborhood}
\alias{neighborhood<-}
\title{Access the neighborhood of a \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}}}
\arguments{
\item{object}{A \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}}}

\item{value}{A neighborhood matrix to replace the existing one}
}
\description{
For details see the documentation of the corresponding class.
}
