% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_network-data.R
\name{network_data}
\alias{network_data}
\title{Create an S4 class that contains \link{node_information} and \link{od_pair_information} for one or multiple networks}
\usage{
network_data(...)
}
\arguments{
\item{...}{objects of of type \link{node_information} and \link{od_pair_information}}
}
\value{
A S4 network data object
}
\description{
Create an S4 class that contains \link{node_information} and \link{od_pair_information} for one or multiple networks
}
