% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{nnodes}
\alias{nnodes}
\title{Access the node count of a \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}} or a
\code{\link[=sp_network_pair-class]{sp_network_pair-class()}}}
\arguments{
\item{object}{One of \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}}, \code{\link[=sp_network_pair-class]{sp_network_pair-class()}}}
}
\description{
For details see the documentation of the corresponding classes.
}
