% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-data.R
\docType{class}
\name{node_information-class}
\alias{node_information-class}
\title{An S4 class to represent information on the nodes of a network.}
\description{
An S4 class to represent information on the nodes of a network.
}
\section{Slots}{

\describe{
\item{\code{network_id}}{A character that serves as identifier for the network}

\item{\code{count}}{A numeric that indicates the number of nodes in the network}

\item{\code{attributes}}{A data.frame that contains all information describing the nodes}

\item{\code{neighborhood}}{A matrix that describes the neighborhood of the nodes}
}}

\seealso{
Other node_information: 
\code{\link{node_information}}
}
\concept{node_information}
