% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-data.R
\name{node_information}
\alias{node_information}
\title{Create an S4 object that contains information in the nodes of a network}
\usage{
node_information(network_id, neighbourhood = NULL, attributes = NULL)
}
\arguments{
\item{network_id}{A character that serves as identifier for the network}

\item{neighbourhood}{A matrix that describes the neighborhood of the nodes}

\item{attributes}{A data.frame that contains all information describing the nodes}
}
\value{
The S4 class node_information
}
\description{
Create an S4 object that contains information in the nodes of a network
}
\seealso{
Other node_information: 
\code{\link{node_information-class}}
}
\concept{node_information}
