% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics_and_maybes.R
\name{npairs}
\alias{npairs}
\title{Generic for accessing the node pairs count of \code{\link[=spflow_network_classes]{spflow_network_classes()}}}
\usage{
npairs(object, ...)
}
\arguments{
\item{object}{A \code{\link[=spflow_network_pairs-class]{spflow_network_pairs-class()}} or \code{\link[=spflow_network_multi-class]{spflow_network_multi-class()}}}

\item{...}{Arguments to be passed to methods}
}
\value{
A numeric, corresponding to the number of node-pairs
}
\description{
For details on the method see the documentation of the
corresponding classes.
}
