% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-data.R
\docType{class}
\name{od_pair_information-class}
\alias{od_pair_information-class}
\title{An S4 class to represent information origin-destination pairs (od-pairs) composed of two nodes.}
\description{
Each origin destination pair is composed of two nodes (see \code{\link[=node_information]{node_information()}}).
All origins belong to the same (origin-) network and all destination belong to
the same (destination-) network.
It is possible to chose the same network for origins and destinations, which
enables to represent od-pairs within the same network.
}
\section{Slots}{

\describe{
\item{\code{origin_id}}{A character that serves as identifier for the origin network}

\item{\code{origin_count}}{A numeric that represents the number of nodes in the origin network}

\item{\code{destination_id}}{A character that serves as identifier for the destination network}

\item{\code{destination_count}}{A numeric that represents the number of nodes in the destination network}

\item{\code{attributes}}{A list of matrices that contain information on node pairs (origins are in rows and destinations are in columns)}
}}

\seealso{
Other od_pair_information network_data: 
\code{\link{od_pair_information}}
}
\concept{od_pair_information network_data}
