% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics_and_maybes.R,
%   R/class_spflow_network_multi.R, R/class_spflow_model.R
\name{pair_corr}
\alias{pair_corr}
\alias{pair_corr,spflow_network_multi-method}
\alias{pair_corr,spflow_model-method}
\title{Correlation matrices for OD data}
\usage{
pair_corr(object, ...)

\S4method{pair_corr}{spflow_network_multi}(
  object,
  id_spflow_network_pairs = id(object)[["pairs"]][[1]],
  spflow_formula,
  add_lags_x = TRUE,
  add_lags_y = FALSE
)

\S4method{pair_corr}{spflow_model}(object, type = "fit", add_fitted = TRUE, add_resid = TRUE, model)
}
\arguments{
\item{object}{A \code{\link[=spflow_model-class]{spflow_model-class()}} or a \code{\link[=spflow_network_multi-class]{spflow_network_multi-class()}}}

\item{...}{Arguments to be passed to methods}

\item{id_spflow_network_pairs}{A character indicating the id of a \code{\link[=spflow_network_pairs-class]{spflow_network_pairs-class()}}}

\item{spflow_formula}{A formula specifying how variables should be used
(for details see section Formula interface in the help page of \code{\link[=spflow]{spflow()}})}

\item{add_lags_x}{A logical, indicating whether spatial lags of the exogenous variables
should be included.}

\item{add_lags_y}{A logical, indicating whether spatial lags of the dependent variables
should be included.}

\item{type}{A character, that should indicate one of \code{c("fit", "empiric")}
\itemize{
\item "fit" will use the moments that have been used during the estimation
\item "empric" will recompute these moments ignoring the weights
}}

\item{add_resid, add_fitted}{A logical, indicating whether the model residuals and fitted value
should be added to the correlation matrix}

\item{model}{A character that should indicate one of \code{paste0("model_", 1:9)}.
The option specifies different correlation structures that are detailed in
the help page of \code{\link[=spflow_control]{spflow_control()}}}
}
\value{
A matrix of pairwise correlations between all variables
}
\description{
The method computes person correlations for all variables available for the
the origins, destinations, and OD-pairs.
The OD-pairs information can be either come from a
\link[=spflow_network_multi-class]{spflow_network_multi} or a
\link[=spflow_model-class]{spflow_models}.
}
\examples{

# Used with a spflow_network_multi ...

# without transformations
corr_mat <- pair_corr(multi_net_usa_ge, "ge_ge")
corr_map(corr_mat)

# with transformations and spatial lags
corr_mat <- pair_corr(
  multi_net_usa_ge,
  "ge_ge",
  y9 ~ . + P_(log(DISTANCE + 1) + .),
  add_lags_y = TRUE)
corr_map(corr_mat)


 # Used with a model...
 gravity_ge <- spflow(
   y1 ~ . + P_(DISTANCE),
   multi_net_usa_ge,
   "ge_ge",
   spflow_control(model = "model_1"))

 corr_mat <- pair_corr(gravity_ge)
 corr_map(corr_mat)

}
\seealso{
\code{\link[=spflow_network_multi-class]{spflow_network_multi-class()}}, \code{\link[=spflow_model-class]{spflow_model-class()}}
}
