% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-list.R
\name{plapply}
\alias{plapply}
\title{A less verbose \code{mapply()}}
\usage{
plapply(..., .f, fix_args = NULL)
}
\arguments{
\item{...}{Named arguments of the same length}

\item{.f}{A function}

\item{fix_args}{A list of named arguments which are held constant}
}
\description{
A less verbose \code{mapply()}
}
\concept{FP}
\keyword{internal}
