% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_model.R
\name{plot,spflow_model-method}
\alias{plot,spflow_model-method}
\title{Diagnostic plots for the results of a spatial interaction model}
\usage{
\S4method{plot}{spflow_model}(x, y, ...)
}
\arguments{
\item{x}{A \code{\link[=spflow_model-class]{spflow_model-class()}}}

\item{y}{not used}

\item{...}{Arguments passed on to other plotting functions}
}
\description{
Diagnostic plots for the results of a spatial interaction model
}
