% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics_and_maybes.R
\name{predict_effect}
\alias{predict_effect}
\title{Generic for predicting the effect of changes in the data}
\usage{
predict_effect(object, ...)
}
\arguments{
\item{object}{A \code{\link[=spflow_model-class]{spflow_model-class()}}}

\item{...}{Arguments to be passed to methods}
}
\value{
The change in outcomes as a matrix, vector or data.frame
}
\description{
For details on the method see the documentation of the
corresponding class.
}
