% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{pull_neighborhood-deprecated}
\alias{pull_neighborhood-deprecated}
\title{Access the neighborhood matrix of the nodes inside a
\code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}
\usage{
pull_neighborhood(object, ...)
}
\arguments{
\item{object}{A \code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}
}
\description{
Access the neighborhood matrix of the nodes inside a
\code{\link[=sp_multi_network-class]{sp_multi_network-class()}}
}
\seealso{
\code{\link{spflow-deprecated}}
}
\keyword{internal}
