% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{pull_neighborhood}
\alias{pull_neighborhood}
\title{Access the neighborhood matrix of the nodes inside a
\code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}
\arguments{
\item{object}{A \code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}
}
\description{
For details see the documentation of the corresponding class.
}
