% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{pull_nodes-deprecated}
\alias{pull_nodes-deprecated}
\title{Access the one or multiple \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}} in a
\code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}
\usage{
pull_nodes(object, ...)
}
\arguments{
\item{object}{A \code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}
}
\description{
For details see the documentation of the corresponding class.
}
\seealso{
\code{\link{spflow-deprecated}}
}
\keyword{internal}
