% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{pull_pairs}
\alias{pull_pairs}
\title{Access the one or multiple \code{\link[=sp_network_pair]{sp_network_pair()}} in a
\code{\link[=sp_multi_network]{sp_multi_network()}}}
\arguments{
\item{object}{A \code{\link[=sp_multi_network]{sp_multi_network()}}}
}
\description{
For details see the documentation of the corresponding class.
}
