% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{results}
\alias{results}
\title{Access results of a \code{\link[=spflow_model-class]{spflow_model-class()}}}
\arguments{
\item{object}{A \code{\link[=spflow_model-class]{spflow_model-class()}}}
}
\description{
For details see the documentation of the corresponding class.
}
