% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics_and_maybes.R
\name{results_flat}
\alias{results_flat}
\title{Access results of a \code{\link[=spflow_model]{spflow_model()}} as a one row data.frame}
\usage{
results_flat(object, ...)
}
\arguments{
\item{object}{spflow_model_meta}

\item{...}{Arguments to be passed to methods}
}
\description{
Access results of a \code{\link[=spflow_model]{spflow_model()}} as a one row data.frame
}
\keyword{internal}
