% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{set_columns-deprecated}
\alias{set_columns-deprecated}
\title{Add columns to the data in a \code{\link[=sp_network_nodes]{sp_network_nodes()}} or
\code{\link[=sp_network_pair]{sp_network_pair()}}}
\usage{
set_columns(object, ...)
}
\arguments{
\item{object}{One of; \code{\link[=sp_network_nodes]{sp_network_nodes()}}, \code{\link[=sp_network_pair]{sp_network_pair()}}}
}
\description{
Add columns to the data in a \code{\link[=sp_network_nodes]{sp_network_nodes()}} or
\code{\link[=sp_network_pair]{sp_network_pair()}}
}
\details{
This function assigns by reference.
}
\seealso{
\code{\link{spflow-deprecated}}
}
\keyword{internal}
