% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sp-network-meta.R
\name{set_columns,sp_network_nodes_pairs-method}
\alias{set_columns,sp_network_nodes_pairs-method}
\title{Internal method that allows to overwrite columns in the data inside
\code{\link[=sp_network_classes]{sp_network_classes()}}}
\usage{
\S4method{set_columns}{sp_network_nodes_pairs}(object, new_cols)
}
\arguments{
\item{new_cols}{New columns to be added to existing data can be one of;
scalar, vector, data.frame}
}
\description{
Internal method that allows to overwrite columns in the data inside
\code{\link[=sp_network_classes]{sp_network_classes()}}
}
\keyword{internal}
