% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_nodes.R
\name{simplfy2df}
\alias{simplfy2df}
\title{Convert spatial data to a simple data.frame}
\usage{
simplfy2df(df, derive_coord_cols = TRUE, prefer_lonlat = TRUE)
}
\description{
Convert spatial data to a simple data.frame
}
\keyword{internal}
