% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sp-network.R
\name{sp_network}
\alias{sp_network}
\title{Create an S4 object that contains information in the nodes of a network}
\usage{
sp_network(
  network_id,
  node_neighborhood = NULL,
  node_data = NULL,
  node_id_column = NULL
)
}
\arguments{
\item{network_id}{A character that serves as identifier for the network}

\item{node_neighborhood}{A matrix that describes the neighborhood of the nodes}

\item{node_data}{A data.frame that contains all information describing the nodes}
}
\value{
The S4 class sp_network
}
\description{
Create an S4 object that contains information in the nodes of a network
}
\concept{sp_network}
