% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sp-network-nodes.R
\name{sp_network_nodes}
\alias{sp_network_nodes}
\title{Create a \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}}}
\usage{
sp_network_nodes(
  network_id,
  node_neighborhood = NULL,
  node_data = NULL,
  node_key_column
)
}
\arguments{
\item{network_id}{A character that serves as identifier for the network}

\item{node_neighborhood}{A matrix that describes the neighborhood of the nodes}

\item{node_data}{A data.frame that contains all information describing the nodes}

\item{node_key_column}{A character indicating the column containing identifiers for the nodes}
}
\value{
The S4 class \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}}
}
\description{
Create a \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}}
}
\examples{
sp_network_nodes("germany",
                 spdep::nb2mat(spdep::poly2nb(germany_grid)),
                 as.data.frame(germany_grid),
                 "NOM")
}
\seealso{
Other Constructors for \link[=sp_network_classes]{spflow network classes}: 
\code{\link{sp_multi_network}},
\code{\link{sp_network_classes}},
\code{\link{sp_network_pair}}
}
\concept{Constructors for \link[=sp_network_classes]{spflow network classes}}
