% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sp-network-pair.R
\name{sp_network_pair}
\alias{sp_network_pair}
\title{Create an S4 object that contains information on origin-destination pairs}
\usage{
sp_network_pair(
  orig_net_id,
  dest_net_id,
  pair_data = NULL,
  orig_key_column,
  dest_key_column
)
}
\arguments{
\item{orig_net_id}{A character that serves as identifier for the origin network}

\item{dest_net_id}{A character that serves as identifier for the destination network}

\item{pair_data}{A data.frame containing information to describe the origin-destination pairs}

\item{orig_key_column}{A character indicating the column containing identifiers for the origins}

\item{dest_key_column}{A character indicating the column containing identifiers for the destinations}
}
\value{
An S4 class of type \code{\link[=sp_network_pair]{sp_network_pair()}}
}
\description{
Create an S4 object that contains information on origin-destination pairs
}
\examples{
pair_frame <- data.table::CJ(origins = germany_grid$NOM,
                             destinations = germany_grid$NOM)
sp_network_pair("ge","ge",pair_frame,"origins","destinations")
}
\seealso{
Other spflow network objects: 
\code{\link{sp_multi_network-class}},
\code{\link{sp_multi_network}},
\code{\link{sp_network_classes}},
\code{\link{sp_network_nodes-class}},
\code{\link{sp_network_nodes}},
\code{\link{sp_network_pair-class}}
}
\concept{spflow network objects}
