% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-simulations.R
\name{spatial_filter}
\alias{spatial_filter}
\title{Derive the spatial filter of an interaction model}
\usage{
spatial_filter(weight_matrices, autoreg_parameters)
}
\arguments{
\item{weight_matrices}{List of flow neighborhood matrices}

\item{autoreg_parameters}{A numeric containing values for the
auto-regressive parameters}
}
\description{
Use the neighborhood matrices of origins and destinations to derive the
three neighborhood matrices of the origin-destination flows.
}
\seealso{
Other spflow simulation functions: 
\code{\link{expand_flow_neighborhood}()},
\code{\link{spflow_sim}()}
}
\concept{spflow simulation functions}
\keyword{internal}
