% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-simulations.R
\name{spatial_model_order}
\alias{spatial_model_order}
\title{Derive the order of the spatial model}
\usage{
spatial_model_order(model = "model_9")
}
\arguments{
\item{model}{A character indicating the model number,  indicating different spatial
dependence structures (see documentation for details), should be one of
\code{paste0("model_", 1:9)}}
}
\value{
An integer corresponding to the number of weight matrices
}
\description{
The order of a spatial model corresponds to the number of neighborhood
matrices that is used to describe the spatial dependence.
}
\keyword{internal}
