% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-package.R
\name{spflow-deprecated}
\alias{spflow-deprecated}
\title{Deprecated functions in package \pkg{spflow}.}
\description{
The functions listed below are deprecated and will be defunct in
the near future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("<function>-deprecated")}.

The functions listed below are defunct. When possible,
alternative functions with similar functionality are also mentioned. Help
pages for deprecated functions are available at
\code{help("<function>-defunct")}.
}
\keyword{internal}
