% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow.R
\name{spflow}
\alias{spflow}
\title{Spatial interaction model}
\usage{
spflow(
  flow_formula,
  sp_multi_network,
  network_pair_id = ids(sp_multi_network, "network_pairs")[[1]],
  flow_control = spflow_control(),
  origin_network_id = NULL,
  destination_network_id = NULL,
  use_sdm = TRUE,
  use_intra = TRUE
)
}
\arguments{
\item{sp_multi_network}{A \code{\link[=sp_multi_network]{sp_multi_network()}} object.}

\item{network_pair_id}{A character indicating the id of a \code{\link[=sp_network_pair]{sp_network_pair()}}}

\item{flow_control}{A \code{\link[=spflow_control]{spflow_control()}} list to fine tune the estimation}

\item{origin_network_id}{A character that serves as identifier for the origin network}

\item{destination_network_id}{A character that serves as identifier for the destination network}

\item{use_sdm}{A logical which adds spatial lags of origin and destination attributes as explanatory variables to the model.}

\item{use_intra}{A logical which adds sperate set of coefficients for intra-observational flows (origin == destination) to the model.}

\item{flow_forumula}{A formulas corresponding to the structural interaction model}
}
\value{
A spflow_model object
}
\description{
Spatial interaction model
}
