% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-control.R
\name{spflow_control}
\alias{spflow_control}
\title{Define details of the estimation procedure with the \code{\link[=spflow]{spflow()}} function.}
\usage{
spflow_control(
  estimation_method = "mle",
  model = "model_9",
  use_intra = TRUE,
  use_sdm = TRUE,
  sdm_variables = "same",
  instrumental_variables = "same",
  weight_variable = NULL,
  decorrelate_instruments = FALSE,
  reduce_pair_instruments = TRUE,
  hessian_method = "mixed"
)
}
\arguments{
\item{estimation_method}{A character which indicates the estimation method, should be one of
\code{c("mle","s2sls","mcmc")}}

\item{model}{A character indicating the model number,  indicating different spatial
dependence structures (see documentation for details), should be one of
\code{paste0("model_", 1:9)}}

\item{use_intra}{A logical which activates the option to use a separate set of parameters
for intra-regional flows (origin == destination)}

\item{use_sdm}{A logical which adds spatial lags of origin and destination attributes as
explanatory variables to the model.}

\item{sdm_variables}{Either a formula or a character; the formula can be used to explicitly
declare the variables in SDM specification, the character should be one of
\code{c("same", "all")} which are short cuts for using all available variables
or the same as used in the main formula provided to \code{\link[=spflow]{spflow()}}}

\item{instrumental_variables}{Either a formula or a character; the formula can be used to explicitly
declare the variables that should be used as instruments during S2SLS
estimation, the character should be one of \code{c("same", "all")} which
are short cuts for using all available variables or the same as used in
the main formula provided to \code{\link[=spflow]{spflow()}}}

\item{weight_variable}{A character indicating the name of a variable that should be used to
weight the origin-destination pairs}

\item{decorrelate_instruments}{A logical whether to perform a PCA to remove (linear) correlation from the
instruments generated for the S2SLS estimator}

\item{reduce_pair_instruments}{A logical that indicates whether the number of instruments that are
derived from pair attributes should be reduced or not (default is TRUE
because constructing these instruments is often the most demanding part of
the estimation.)}

\item{hessian_method}{A character which indicates the method for hessian calculation}
}
\value{
A list of control parameters for estimation via \code{\link[=spflow]{spflow()}}
}
\description{
This function creates a list to fine tune the estimation of a spatial
interaction model with \code{\link[=spflow]{spflow()}}.
The options allow to adjust the estimation method and give the user full
control over the use of the explanatory variables.
The user can also adjust the form of autocorrelation to be considered.
}
\section{Adjusting the form of autocorrelation}{


The option \code{model} allows to declare one of nine different forms of
autocorrelation that follow the naming convention of
\insertCite{LeSage2008;textual}{spflow}.
The most general specification is "model_9" and all other correspond to
special cases of this one.
The different sub models are summarized in this table.
}

\examples{

# default is MLE estimation of the most comprehensive model
default_control <- spflow_control()

# change the estimation method
custom_control <- spflow_control(estimation_method = "mcmc")

# change the form of autocorrelation to be considered
custom_control <- spflow_control(model = "model_7")

# declare precisely which variables are to be used in the SDM form
custom_control <-
  spflow_control(sdm_variables = ~ O_(v1 + v2) + D_(v2 + v3) + I_(v1 + v4))

# deactivate the intra-regional coefficients and SDM variables
custom_control <- spflow_control(use_intra = FALSE, use_sdm = FALSE)
custom_control <- spflow_control(use_intra = FALSE, sdm_variables = "none")
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=spflow]{spflow()}}
}
