% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics_and_maybes.R
\name{spflow_generics}
\alias{spflow_generics}
\alias{complete_pairs}
\alias{coord}
\alias{dat}
\alias{dat<-}
\alias{dat-set}
\alias{id}
\alias{id<-}
\alias{id-set}
\alias{mcmc_results}
\alias{neighborhood}
\alias{neighborhood<-}
\alias{neighborhood-set}
\alias{nnodes}
\alias{npairs}
\alias{pull_member}
\alias{results}
\alias{results<-}
\alias{results_flat}
\alias{sd_error}
\alias{update_dat}
\alias{varcov}
\title{Generic functions for the \code{\link[=spflow_network_classes]{spflow_network_classes()}}
and the \code{\link[=spflow_model-class]{spflow_model-class()}}}
\usage{
complete_pairs(object, ...)

coord(object, ...) # internal

dat(object, ...)

dat(object, ...) <- value

id(object, ...)

id(object, ...) <- value

mcmc_results(object)

neighborhood(object, ...)

neighborhood(object) <- value

nnodes(object, ...)

npairs(object, ...)

pull_member(object, ...)

results(object, ...)

results(object) <- value # internal

results_flat(object, ...)

sd_error(object)

update_dat(object, ...)

varcov(object)
}
\arguments{
\item{object}{A \code{\link[=spflow_model-class]{spflow_model-class()}} or \code{\link[=spflow_network_classes]{spflow_network_classes()}}}

\item{...}{Arguments passed on to methods}

\item{value}{An appropriate replacement value}
}
\description{
Most of these generic functions link to simple accessing or
replacement methods, and do not have their own documentation page.
The associated methods are documented within the corresponding classes.

Generics with more complex methods have their dedicated documentation pages.

Also note that some generic functions are currently not exported and meant
to be used by package developers only.
}
\seealso{
\itemize{
\item Documentation of classes: \code{\link[=spflow_model-class]{spflow_model-class()}}, \code{\link[=spflow_network_classes]{spflow_network_classes()}}
\item Other generics and methods: \code{\link[=pair_cor]{pair_cor()}}, \code{\link[=pair_merge]{pair_merge()}}, \code{\link[=predict]{predict()}}, \code{\link[=predict_effect]{predict_effect()}}, \code{\link[=spflow_moran_plots]{spflow_moran_plots()}}, \code{\link[=spflow_map]{spflow_map()}}
}
}
\keyword{internal}
