% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics_and_maybes.R,
%   R/class_spflow_multinet.R, R/class_spflow_model.R
\name{spflow_map}
\alias{spflow_map}
\alias{spflow_map,spflow_multinet-method}
\alias{spflow_map,spflow_model-method}
\title{Geographic representation of flows}
\usage{
spflow_map(object, ...)

\S4method{spflow_map}{spflow_multinet}(object, id_spflow_pairs = id(object)[["pairs"]][[1]], ..., flow_var)

\S4method{spflow_map}{spflow_model}(object, ..., flow_type = "resid", add_title = TRUE)
}
\arguments{
\item{object}{A \code{\link[=spflow_model-class]{spflow_model-class()}}}

\item{...}{further arguments passed to \code{\link[=map_flows]{map_flows()}}}

\item{id_spflow_pairs}{A character indicating the id of a \code{\link[=spflow_pairs]{spflow_pairs()}} (only relevant if
the \code{\link[=spflow_multinet]{spflow_multinet()}} contains multiple \code{spflow_pairs}-objects:
defaults to the of them)}

\item{flow_var}{A character, indicating one variable from the network pair data}

\item{flow_type}{A character indicating what to values to show.
Should be one of c("resid", "fitted", "actual").}

\item{add_title}{A logical, if \code{TRUE} the flow_type is added as title.}
}
\description{
Geographic representation of flows
}
\examples{

 # Used with a spflow_multinet ...
 spflow_map(multi_net_usa_ge, "ge_ge",flow_var = "y9")


 # Used with a model...
 gravity_ge <- spflow(
   y1 ~ . + P_(DISTANCE),
   multi_net_usa_ge,
   "ge_ge",
   spflow_control(model = "model_1"))
 spflow_map(gravity_ge)
 spflow_map(gravity_ge, flow_type = "fitted")
 spflow_map(gravity_ge, flow_type = "actual")

}
\seealso{
\code{\link[=map_flows]{map_flows()}}
}
