% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-spflow-mcmc.R
\name{spflow_mcmc}
\alias{spflow_mcmc}
\title{Moment based estimation procedure for the MCMC estimator}
\usage{
spflow_mcmc(
  ZZ,
  ZY,
  TSS,
  N,
  n_d,
  n_o,
  OW_traces,
  DW_traces,
  flow_control = flow_control,
  nb_draw = 5500,
  nb_burn_in = 2500
)
}
\description{
Moment based estimation procedure for the MCMC estimator
}
\keyword{internal}
