% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_model.R
\docType{class}
\name{spflow_model-class}
\alias{spflow_model-class}
\alias{spflow_model_mcmc,}
\alias{spflow_model_mle,}
\alias{spflow_model_s2sls,}
\alias{spflow_model_ols}
\alias{coef,spflow_model-method}
\alias{fitted,spflow_model-method}
\alias{logLik,spflow_model-method}
\alias{mcmc_results,spflow_model_mcmc-method}
\alias{nobs,spflow_model-method}
\alias{neighborhood,spflow_model-method}
\alias{resid,spflow_model-method}
\alias{results,spflow_model-method}
\alias{results_flat,spflow_model-method}
\alias{sd_error,spflow_model-method}
\alias{varcov,spflow_model_varcov-method}
\title{Class spflow_model}
\usage{
\S4method{coef}{spflow_model}(object, param_subset = NULL)

\S4method{fitted}{spflow_model}(object, return_type = "V")

\S4method{logLik}{spflow_model}(object)

\S4method{mcmc_results}{spflow_model_mcmc}(object)

\S4method{nobs}{spflow_model}(object, which = "sample")

\S4method{neighborhood}{spflow_model}(object, which_nb)

\S4method{resid}{spflow_model}(object, return_type = "V")

\S4method{results}{spflow_model}(object)

\S4method{results_flat}{spflow_model}(
  object,
  coef_info = c("est", "sd"),
  main_info = c("estimation_method", "model_coherence", "R2_corr", "ll", "sd_error")
)

\S4method{sd_error}{spflow_model}(object)

\S4method{varcov}{spflow_model_varcov}(object)
}
\arguments{
\item{object}{A spflow_model}

\item{param_subset}{A character indicating the subset of model parameters to be returned "rho"
relates to the autoregression parameters and "delta" to those of the
exogenous variables.}

\item{return_type}{A character indicating the format of the returned values:
\itemize{
\item "V" leads to an atomic vector
\item "M" leads to a OD matrix where missing data is replaced by zeros
\item "OD" leads to a data.frame with columns being the the values
and the id's of the destinations and the origins
}}

\item{which}{A character vector indicating the subset of observations to consider
should be one of \code{c("fit", "cart", "pop", "pair", "orig", "dest")}.}

\item{which_nb}{A character vector: "OW" for origin- and "DW" for destination neighborhood}

\item{coef_info}{A character indicating column names in the results}

\item{main_info}{A character indicating named elements in the estimation_control or estimation_diagnostics}
}
\description{
An S4 class that contains the estimation results of spatial econometric
interaction models estimated by the \code{\link[=spflow]{spflow()}} function.

There are four subclasses that are specific to the chosen estimation method
(OLS, MLE, Bayesian MCMC or S2SLS).
They contain some additional information specific to the corresponding
method but most behaviors and data are identical among them.
}
\section{Slots}{

\describe{
\item{\code{estimation_results}}{A data.frame that contains the main \code{\link[=results]{results()}} of the estimation}

\item{\code{estimation_control}}{A list that contains all control parameters of the estimation
(see \code{\link[=spflow_control]{spflow_control()}})}

\item{\code{estimation_diagnostics}}{A list of further indicators about the estimation}

\item{\code{spflow_formula}}{A formula}

\item{\code{spflow_networks}}{A \code{\link[=spflow_network_multi-class]{spflow_network_multi-class()}}}

\item{\code{spflow_matrices}}{A list or NULL}

\item{\code{spflow_formula}}{The formula used to fit the model}

\item{\code{spflow_indicators}}{A data.frame containing the indicators of od-pairs}

\item{\code{spflow_moments}}{A list of moment matrices used for estimating the model}

\item{\code{spflow_nbfunctions}}{A list that may contain a function to calculate the log-determinant term
and one to validate the parameter space for the spatial interaction model.}
}}

\section{Main results}{

The main results are accessed with the \code{results()} method.
They are given in the form of a data frame with the following columns;
\itemize{
\item \code{est}: value of the estimated parameter
\item \code{sd}: value of the standard deviation of the parameter
\item \code{t.test}: value of the t-statistic under the two-sided hypothesis that
the parameter value is 0.
\item \code{p.val}: the p-value associated to the t-test
\item \code{quant_025}: for Bayesian estimation the lower bound of 95\% interval
\item \code{quant_975}: for Bayesian estimation the upper bound of 95\% interval
}
}

\examples{

spflow_results <- spflow(y9 ~ . + P_(DISTANCE), multi_net_usa_ge)

# General methods
results(spflow_results) # data.frame of main results
coef(spflow_results) # vector of estimated coefficients
fitted(spflow_results) # vector of fitted values
resid(spflow_results) # vector of residuals
nobs(spflow_results) # number of observations
sd_error(spflow_results) # standard deviation of the error term
predict(spflow_results) # computation of the in sample predictor
plot(spflow_results) # some plots for assessing the model

# MLE methods
logLik(spflow_results) # value of the likelihood function

# MLE, OLS and S2SLS methods
varcov(spflow_results) # variance covariance matrix of the estimators

# MCMC methods
spflow_results_mcmc <- spflow(
  y2 ~ . + P_(DISTANCE),
  multi_net_usa_ge,
  estimation_control = spflow_control(estimation_method = "mcmc",
                                model = "model_2"))
results(spflow_results_mcmc)
plot(mcmc_results(spflow_results_mcmc)) # parameter values during the mcmc sampling
}
\seealso{
\code{\link[=spflow]{spflow()}}, \code{\link[=spflow_network_classes]{spflow_network_classes()}}
}
\author{
Lukas Dargel
}
