% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow-model-meta.R
\name{spflow_model}
\alias{spflow_model}
\title{Internal function to construct a \code{\link[=spflow_model-class]{spflow_model-class()}}}
\usage{
spflow_model(
  ...,
  estimation_results,
  estimation_control,
  N,
  sd_error,
  R2_corr = NULL,
  resid = NULL,
  fitted = NULL,
  spatial_filter_matrix = NULL,
  design_matrix = NULL,
  model_moments = NULL
)
}
\arguments{
\item{...}{Further arguments passed to more specific classes in accordance to the
estimation method}

\item{estimation_results}{A data.frame of estimation \code{\link[=results]{results()}}}

\item{estimation_control}{A list of control parameters}

\item{N}{A numeric indicating the number of observations}

\item{sd_error}{A numeric which reports the}

\item{R2_corr}{A numeric which reports a pseudo R^2 measure}

\item{resid}{A numeric vector of regression residuals}

\item{fitted}{A numeric vector of fitted values}

\item{spatial_filter_matrix}{A matrix which represents the spatial filter}

\item{design_matrix}{The design matrix/matrices of the model}
}
\description{
Internal function to construct a \code{\link[=spflow_model-class]{spflow_model-class()}}
}
\keyword{internal}
