% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-model-matrix_1.R
\name{spflow_model_matrix}
\alias{spflow_model_matrix}
\title{Generate design matrices for the spatial interaction model}
\usage{
spflow_model_matrix(
  sp_multi_network,
  network_pair_id,
  flow_formula,
  flow_control
)
}
\arguments{
\item{sp_multi_network}{A \code{\link[=sp_multi_network]{sp_multi_network()}} object that contains information on the origins,
and the destinations and their neighborhood structure}

\item{network_pair_id}{A character indicating the id of a \code{\link[=sp_network_pair]{sp_network_pair()}}, defaults to the
first pair inside the \code{\link[=sp_multi_network]{sp_multi_network()}}}

\item{flow_formula}{A formula specifying the spatial interaction model (for details see
section Formula interface)}

\item{flow_control}{A list generated by \code{\link[=spflow_control]{spflow_control()}} that provides fine grained control
over the estimation procedure}
}
\value{
A list of design matrices for spatial interaction model
}
\description{
Creates model matrices from a \code{\link[=sp_multi_network]{sp_multi_network()}} network.
For efficiency the relational representation of origin and destination data
is preserved.
}
\keyword{internal}
