% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-model-moments_1.R
\name{spflow_model_moments}
\alias{spflow_model_moments}
\title{Compute the principal moments for a spatial flow model}
\usage{
spflow_model_moments(...)
}
\description{
Compute the principal moments for a spatial flow model
}
\keyword{internal}
