% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics_and_maybes.R,
%   R/class_spflow_multinet.R, R/class_spflow_model.R
\name{spflow_moran_plots}
\alias{spflow_moran_plots}
\alias{spflow_moran_plots,spflow_multinet-method}
\alias{spflow_moran_plots,spflow_model-method}
\title{Moran scatter plots of interaction data}
\usage{
spflow_moran_plots(object, ...)

\S4method{spflow_moran_plots}{spflow_multinet}(
  object,
  id_spflow_pairs = id(object)[["pairs"]][[1]],
  flow_var,
  model = "model_9",
  DW,
  OW,
  add_lines = TRUE
)

\S4method{spflow_moran_plots}{spflow_model}(object, model = "model_9", DW, OW, add_lines = TRUE)
}
\arguments{
\item{object}{A \code{\link[=spflow_multinet-class]{spflow_multinet-class()}} or a \code{\link[=spflow_model-class]{spflow_model-class()}}}

\item{...}{arguments passed to methods}

\item{id_spflow_pairs}{A character indicating the id of a \code{\link[=spflow_pairs]{spflow_pairs()}} (only relevant if
the \code{\link[=spflow_multinet]{spflow_multinet()}} contains multiple \code{spflow_pairs}-objects:
defaults to the of them)}

\item{flow_var}{A character, indicating one variable from the network pair data}

\item{model}{A character indicating the model number,  indicating different spatial
dependence structures (see documentation for details), should be one of
\code{paste0("model_", 1:9)}}

\item{DW, OW}{A matrix to replace the neighborhood of the destinations (DW) and origins (OW).
Defaults to the one supplied to the model.}

\item{add_lines}{A logical, if \code{TRUE} regression lines are added to the Moran scatter plots.}
}
\description{
Moran scatter plots of interaction data
}
\examples{

 # Used with a spflow_multinet ...
 # To check the if there is spatial correlation in any variable
 spflow_moran_plots(multi_net_usa_ge, "ge_ge",flow_var = "y9")


 # Used with a model...
 # To check the if there is spatial correlation in the residuals
 gravity_ge <- spflow(
   y9 ~ . + P_(DISTANCE),
   multi_net_usa_ge,
   "ge_ge",
   spflow_control(model = "model_1"))
 spflow_moran_plots(gravity_ge)

}
