% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_multinet.R
\name{spflow_multinet}
\alias{spflow_multinet}
\title{Create an S4 class that contains \code{\link[=spflow_nodes]{spflow_nodes()}} and \code{\link[=spflow_pairs]{spflow_pairs()}} for one or multiple nodes}
\usage{
spflow_multinet(...)
}
\arguments{
\item{...}{objects of type \code{\link[=spflow_nodes]{spflow_nodes()}} and \code{\link[=spflow_pairs]{spflow_pairs()}}}
}
\value{
An S4 class of type \code{\link[=spflow_multinet-class]{spflow_multinet-class()}}
}
\description{
Create an S4 class that contains \code{\link[=spflow_nodes]{spflow_nodes()}} and \code{\link[=spflow_pairs]{spflow_pairs()}} for one or multiple nodes
}
\examples{
spflow_multinet() # empty
spflow_multinet(germany_net,usa_net) # two nodes, no pairs
}
\seealso{
Other Constructors for spflow network classes: 
\code{\link{spflow_network_classes}},
\code{\link{spflow_nodes}},
\code{\link{spflow_pairs}}
}
\concept{Constructors for spflow network classes}
