% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_network.R
\docType{class}
\name{spflow_network-class}
\alias{spflow_network-class}
\alias{dat,spflow_network-method}
\alias{dat<-,spflow_network-method}
\alias{id,spflow_network-method}
\alias{id<-,spflow_network-method}
\alias{neighborhood,spflow_network-method}
\alias{neighborhood<-,spflow_network-method}
\alias{nnodes,spflow_network-method}
\alias{update_dat,spflow_network-method}
\title{spflow_network Class}
\usage{
\S4method{dat}{spflow_network}(object)

\S4method{dat}{spflow_network}(
  object,
  node_key_column,
  node_coord_columns,
  derive_coordinates = FALSE,
  prefer_lonlat = TRUE
) <- value

\S4method{id}{spflow_network}(object)

\S4method{id}{spflow_network}(object) <- value

\S4method{neighborhood}{spflow_network}(object)

\S4method{neighborhood}{spflow_network}(object) <- value

\S4method{nnodes}{spflow_network}(object)

\S4method{update_dat}{spflow_network}(object, new_dat)
}
\arguments{
\item{object}{A spflow_network-class}

\item{value}{An object to replace the existing id/data/neighborhood}

\item{node_key_column}{A character indicating the column containing the identifiers for the nodes}

\item{node_coord_columns}{A character indicating the columns that represent the coordinates of the
nodes. For example \code{c("LON", "LAT")}.}

\item{derive_coordinates}{A logical indicating whether there should be an attempt to infer the
coordinates from the node_data.}

\item{prefer_lonlat}{A logical indicating whether the coordinates should be transformed to
longitude and latitude.}

\item{new_dat}{A data.frame}
}
\description{
An S4 class that contains all information on a single network.
In this representation a network is composed of nodes which must be
identified uniquely by and ID.
Each node is described by variables stored in a data.frame.
The node neighborhood matrix describes strength of links between the nodes
of the network.
The class is constructed by the \code{\link[=spflow_network]{spflow_network()}} function.
}
\section{Slots}{

\describe{
\item{\code{id_net}}{A character that serves as an identifier for the set of nodes}

\item{\code{node_data}}{A data.frame that contains all information describing the nodes}

\item{\code{node_neighborhood}}{A matrix that describes the neighborhood relations of the nodes}
}}

\examples{
## access the data describing the nodes
new_dat <- dat(germany_net)

# access the id of the network
germany_net2 <- germany_net
id(germany_net2)
id(germany_net2) <- "Germany"

# access the neighborhood matrix of the nodes
neighborhood(germany_net)

# access the number of nodes inside the network
nnodes(germany_net)

}
