% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_network.R
\name{spflow_network}
\alias{spflow_network}
\title{Create a \code{\link[=spflow_network-class]{spflow_network-class()}}}
\usage{
spflow_network(
  id_net,
  node_neighborhood = NULL,
  node_data = NULL,
  node_key_column,
  node_coord_columns,
  derive_coordinates = missing(node_coord_columns),
  prefer_lonlat = TRUE,
  normalize_byrow = FALSE
)
}
\arguments{
\item{id_net}{A character that serves as an identifier for the set of nodes}

\item{node_neighborhood}{A matrix that describes the neighborhood of the nodes}

\item{node_data}{A data.frame that contains all information describing the nodes}

\item{node_key_column}{A character indicating the column containing the identifiers for the nodes}

\item{node_coord_columns}{A character indicating the columns that represent the coordinates of the
nodes. For example \code{c("LON", "LAT")}.}

\item{derive_coordinates}{A logical indicating whether there should be an attempt to infer the
coordinates from the node_data.}

\item{prefer_lonlat}{A logical indicating whether the coordinates should be transformed to
longitude and latitude.}

\item{normalize_byrow}{A logical, if \code{TRUE} the neighborhood will be row-normalized, otherwise
it is scaled to have a spectral radius of one.}
}
\value{
An S4 class of type \code{\link[=spflow_network-class]{spflow_network-class()}}
}
\description{
Create a \code{\link[=spflow_network-class]{spflow_network-class()}}
}
\examples{
spflow_network(
  "germany",
  spdep::nb2mat(spdep::poly2nb(germany_grid)),
  as.data.frame(germany_grid),
  "ID_STATE")
}
