% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_network_nodes.R
\docType{class}
\name{spflow_network_nodes-class}
\alias{spflow_network_nodes-class}
\alias{dat,spflow_network_nodes-method}
\alias{dat<-,spflow_network_nodes-method}
\alias{id,spflow_network_nodes-method}
\alias{id<-,spflow_network_nodes-method}
\alias{neighborhood,spflow_network_nodes-method}
\alias{neighborhood<-,spflow_network_nodes-method}
\alias{nnodes,spflow_network_nodes-method}
\alias{update_dat,spflow_network_nodes-method}
\title{spflow_network_nodes Class}
\usage{
\S4method{dat}{spflow_network_nodes}(object)

\S4method{dat}{spflow_network_nodes}(object) <- value

\S4method{id}{spflow_network_nodes}(object)

\S4method{id}{spflow_network_nodes}(object) <- value

\S4method{neighborhood}{spflow_network_nodes}(object)

\S4method{neighborhood}{spflow_network_nodes}(object) <- value

\S4method{nnodes}{spflow_network_nodes}(object)

\S4method{update_dat}{spflow_network_nodes}(object, new_dat)
}
\arguments{
\item{object}{A spflow_network_nodes-class}

\item{value}{An object to replace the existing id/data/neighborhood}

\item{new_dat}{A data.frame}
}
\description{
An S4 class that contains all information on a single network.
In this representation a network is composed of nodes which must be
identified uniquely by and ID.
Each node is described by variables stored in a data.frame.
The node neighborhood matrix describes strength of links between the nodes
of the network.
The class is constructed by the \code{\link[=spflow_network_nodes]{spflow_network_nodes()}} function.
}
\section{Slots}{

\describe{
\item{\code{id_spflow_network_nodes}}{A character that serves as an identifier for the set of nodes}

\item{\code{node_data}}{A data.frame that contains all information describing the nodes}

\item{\code{node_neighborhood}}{A matrix that describes the neighborhood relations of the nodes}
}}

\examples{
## access the data describing the nodes
new_dat <- dat(germany_net)

# access the id of the network
germany_net2 <- germany_net
id(germany_net2)
id(germany_net2) <- "Germany"

# access the neighborhood matrix of the nodes
neighborhood(germany_net)

# access the number of nodes inside the network
nnodes(germany_net)

}
\seealso{
Other spflow network classes: 
\code{\link{spflow_network_classes}},
\code{\link{spflow_network_multi-class}},
\code{\link{spflow_network_pairs-class}}
}
\concept{spflow network classes}
