% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_network_pair.R
\name{spflow_network_pair}
\alias{spflow_network_pair}
\title{Create a \code{\link[=spflow_network_pair-class]{spflow_network_pair-class()}}}
\usage{
spflow_network_pair(
  id_orig_net,
  id_dest_net,
  id_net_pair = paste0(id_orig_net, "_", id_dest_net),
  pair_data = NULL,
  orig_key_column,
  dest_key_column
)
}
\arguments{
\item{id_orig_net}{A character that serves as identifier for the origin network}

\item{id_dest_net}{A character that serves as identifier for the destination network}

\item{id_net_pair}{A character that as identifier for network_pair}

\item{pair_data}{A data.frame containing information on the origin-destination pairs}

\item{orig_key_column}{A character indicating the name of the column containing the identifiers
of the origins}

\item{dest_key_column}{A character indicating the name of the column containing the identifiers
of the destinations}
}
\value{
A \code{\link[=spflow_network_pair-class]{spflow_network_pair-class()}}
}
\description{
Create a \code{\link[=spflow_network_pair-class]{spflow_network_pair-class()}}
}
\examples{
pair_frame <- data.frame(
  ORIG_ID_STATE = rep(germany_grid$ID_STATE, times = 16),
  DEST_ID_STATE = rep(germany_grid$ID_STATE, each = 16))
spflow_network_pair("ge","ge","ge_ge",pair_frame,"ORIG_ID_STATE","DEST_ID_STATE")
}
