% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_network_pairs.R
\name{spflow_network_pairs}
\alias{spflow_network_pairs}
\title{Constructor for the \code{\link[=spflow_network_pairs-class]{spflow_network_pairs-class()}}}
\usage{
spflow_network_pairs(
  id_orig_nodes,
  id_dest_nodes,
  id_spflow_network_pairs = paste0(id_orig_nodes, "_", id_dest_nodes),
  pair_data = NULL,
  orig_key_column,
  dest_key_column
)
}
\arguments{
\item{id_orig_nodes}{A character that serves as identifier for the origin network}

\item{id_dest_nodes}{A character that serves as identifier for the destination network}

\item{id_spflow_network_pairs}{A character that as identifier for network_pair}

\item{pair_data}{A data.frame containing information on the origin-destination pairs}

\item{orig_key_column}{A character indicating the name of the column containing the identifiers
of the origins}

\item{dest_key_column}{A character indicating the name of the column containing the identifiers
of the destinations}
}
\value{
A \code{\link[=spflow_network_pairs-class]{spflow_network_pairs-class()}}
}
\description{
Constructor for the \code{\link[=spflow_network_pairs-class]{spflow_network_pairs-class()}}
}
\examples{
pair_frame <- data.frame(
  ORIG_ID_STATE = rep(germany_grid$ID_STATE, times = 16),
  DEST_ID_STATE = rep(germany_grid$ID_STATE, each = 16))
spflow_network_pairs("ge","ge","ge_ge",pair_frame,"ORIG_ID_STATE","DEST_ID_STATE")
}
