% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_nodes.R
\name{spflow_nodes}
\alias{spflow_nodes}
\title{Create a \code{\link[=spflow_nodes-class]{spflow_nodes-class()}}}
\usage{
spflow_nodes(
  network_id,
  node_neighborhood = NULL,
  node_data = NULL,
  node_key_column,
  node_coord_columns,
  derive_coordinates = missing(node_coord_columns),
  prefer_lonlat = TRUE,
  normalize_neighborhood = FALSE,
  normalize_byrow = TRUE
)
}
\arguments{
\item{network_id}{A character that serves as an identifier for the network}

\item{node_neighborhood}{A matrix that describes the neighborhood of the nodes}

\item{node_data}{A data.frame that contains all information describing the nodes}

\item{node_key_column}{A character indicating the column containing the identifiers for the nodes}

\item{node_coord_columns}{A character indicating the columns that represent the coordinates of the
nodes. For example \code{c("LON", "LAT")}.}

\item{derive_coordinates}{A logical indicating whether there should be an attempt to infer the
coordinates from the node_data.}

\item{prefer_lonlat}{A logical indicating whether the coordinates should be transformed to
longitude and latitude.}

\item{normalize_neighborhood}{A logical indicating whether the node_neighborhood matrix should be
normalized to satisfy the usual assumptions in spatial econometrics.
This means zeros on the main diagonal and a spectral radius of one.}

\item{normalize_byrow}{A logical, if \code{TRUE} the neighborhood is row-normalized.}
}
\value{
An S4 class of type \code{\link[=spflow_nodes-class]{spflow_nodes-class()}}
}
\description{
Create a \code{\link[=spflow_nodes-class]{spflow_nodes-class()}}
}
\examples{
spflow_nodes("germany",
                 spdep::nb2mat(spdep::poly2nb(germany_grid)),
                 as.data.frame(germany_grid),
                 "ID_STATE")
}
\seealso{
Other Constructors for spflow network classes: 
\code{\link{spflow_multinet}},
\code{\link{spflow_network_classes}},
\code{\link{spflow_pairs}}
}
\concept{Constructors for spflow network classes}
