% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_pairs.R
\docType{class}
\name{spflow_pairs-class}
\alias{spflow_pairs-class}
\alias{dat,spflow_pairs-method}
\alias{dat<-,spflow_pairs-method}
\alias{id,spflow_pairs-method}
\alias{id<-,spflow_pairs-method}
\alias{npairs,spflow_pairs-method}
\alias{nnodes,spflow_pairs-method}
\title{Class spflow_pairs}
\usage{
\S4method{dat}{spflow_pairs}(object)

\S4method{dat}{spflow_pairs}(object) <- value

\S4method{id}{spflow_pairs}(object)

\S4method{id}{spflow_pairs}(object, which = "pair") <- value

\S4method{npairs}{spflow_pairs}(object)

\S4method{nnodes}{spflow_pairs}(object)
}
\arguments{
\item{object}{A spflow_pairs-class}

\item{value}{An object to replace the existing id/data}

\item{which}{A character indicating which of the ids to change, should be one of
\code{c("origin", "destination", "pair")}.
The first characters may be used as shortcuts.}
}
\description{
An S4 class which holds information on origin-destination (OD) pairs.
Each OD pair is composed of two nodes (see \code{\link[=spflow_nodes-class]{spflow_nodes-class()}}).
All origins belong to the same (origin-) network and all destination belong
to the same (destination-) network. It is possible to choose the same
network for origins and destinations, which enables to represent OD pairs
within the same network.
}
\section{Slots}{

\describe{
\item{\code{orig_net_id}}{A character that serves as identifier for the origin network}

\item{\code{dest_net_id}}{A character that serves as identifier for the destination network}

\item{\code{network_pair_id}}{A character identifying the pair of networks}

\item{\code{pair_data}}{A data.frame containing information on origin-destination pairs}
}}

\examples{
## access the data describing the node pairs

net_pair_ge_ge <- pull_member(multi_net_usa_ge,"ge_ge")
dat(net_pair_ge_ge)

## access the id of a network pair

net_pair_ge_ge <- pull_member(multi_net_usa_ge,"ge_ge")
id(net_pair_ge_ge)
id(net_pair_ge_ge) <- "Germany_Germany"

## access the number of node pairs in a network pair

net_pair_ge_ge <- pull_member(multi_net_usa_ge,"ge_ge")
npairs(net_pair_ge_ge)

## access the number of origin and destination nodes in a network pair
net_pair_ge_ge <- pull_member(multi_net_usa_ge,"ge_ge")
nnodes(net_pair_ge_ge)
nnodes(net_pair_ge_ge)["orig"]
nnodes(net_pair_ge_ge)["dest"]
prod(nnodes(net_pair_ge_ge) == npairs(net_pair_ge_ge))

}
\seealso{
Other spflow network classes: 
\code{\link{spflow_multinet-class}},
\code{\link{spflow_network_classes}},
\code{\link{spflow_nodes-class}}
}
\concept{spflow network classes}
