% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_pairs.R
\name{spflow_pairs}
\alias{spflow_pairs}
\title{Create an S4 object that contains information on origin-destination pairs}
\usage{
spflow_pairs(
  orig_net_id,
  dest_net_id,
  network_pair_id = paste0(orig_net_id, "_", dest_net_id),
  pair_data = NULL,
  orig_key_column,
  dest_key_column
)
}
\arguments{
\item{orig_net_id}{A character that serves as identifier for the origin network}

\item{dest_net_id}{A character that serves as identifier for the destination network}

\item{network_pair_id}{A character that as identifier for network_pair}

\item{pair_data}{A data.frame containing information on the origin-destination pairs}

\item{orig_key_column}{A character indicating the name of the column containing the identifiers
of the origins}

\item{dest_key_column}{A character indicating the name of the column containing the identifiers
of the destinations}
}
\value{
An S4 class of type \code{\link[=spflow_pairs-class]{spflow_pairs-class()}}
}
\description{
Create an S4 object that contains information on origin-destination pairs
}
\examples{
pair_frame <- data.frame(
  ORIG_ID_STATE = rep(germany_grid$ID_STATE, times = 16),
  DEST_ID_STATE = rep(germany_grid$ID_STATE, each = 16))
spflow_pairs("ge","ge","ge_ge",pair_frame,"ORIG_ID_STATE","DEST_ID_STATE")
}
\seealso{
Other Constructors for spflow network classes: 
\code{\link{spflow_multinet}},
\code{\link{spflow_network_classes}},
\code{\link{spflow_nodes}}
}
\concept{Constructors for spflow network classes}
