% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow_estimation.R
\name{spflow_post_estimation}
\alias{spflow_post_estimation}
\title{Internal method to add details to a \code{\link[=spflow_model-class]{spflow_model-class()}}}
\usage{
spflow_post_estimation(
  object,
  spflow_data,
  spflow_indicators,
  spflow_moments,
  spflow_neighborhood,
  spflow_matrices
)
}
\value{
An object of class spflow_model
}
\description{
Internal method to add details to a \code{\link[=spflow_model-class]{spflow_model-class()}}
}
\details{
The method adds the design matrix and the coefficient names to an
\code{\link[=spflow_model-class]{spflow_model-class()}} object.
It also calculates the fitted values and the residuals as well as a
goodness-of-fit measure.
}
\keyword{internal}
