% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow_simulations.R
\name{spflow_sim}
\alias{spflow_sim}
\title{Simulate spatial interactions}
\usage{
spflow_sim(
  exogenous_variables,
  model_coefficients,
  filter_matrix,
  noise_sd,
  verbose = FALSE
)
}
\arguments{
\item{exogenous_variables}{A matrix of exogenous variables}

\item{model_coefficients}{A numeric vector of coefficients}

\item{filter_matrix}{A matrix that represents the \code{\link[=spatial_filter]{spatial_filter()}} of an interaction model}

\item{noise_sd}{A numeric which indicates the desired standard deviation of the simulated noise}

\item{verbose}{A logical whether signal to noise ration should be printed}
}
\value{
A vector of simulated flows
}
\description{
Simulate spatial interactions
}
\seealso{
Other spflow simulation functions: 
\code{\link{expand_spflow_neighborhood}()},
\code{\link{spatial_filter}()}
}
\concept{spflow simulation functions}
\keyword{internal}
