% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-logdet.R
\name{trace_lookup_template}
\alias{trace_lookup_template}
\title{Create a lookup table for the calculation of traces}
\usage{
trace_lookup_template(aprox_order)
}
\arguments{
\item{aprox_order}{An integer defining the order of the taylor series}
}
\value{
A data.table that can serve as a lookup
}
\description{
Create a lookup table for the calculation of traces
}
\keyword{internal}
