% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics_and_maybes.R
\name{update_dat}
\alias{update_dat}
\title{Generic for updating data related to specific rows in
\code{\link[=spflow_network_classes]{spflow_network_classes()}}}
\usage{
update_dat(object, new_dat)
}
\arguments{
\item{object}{One among \code{\link[=spflow_network_classes]{spflow_network_classes()}}}
}
\value{
The updated object
}
\description{
For details on the methods see the documentation of the
corresponding classes.
}
