% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{varcov}
\alias{varcov}
\title{Generic for accessing the variance-covariance matrix of the
parameters inside a \code{\link[=spflow_model-class]{spflow_model-class()}}.}
\usage{
varcov(object)
}
\arguments{
\item{object}{A \code{\link[=spflow_model-class]{spflow_model-class()}}}
}
\value{
The variance-covariance matrix of the fitted model parameters
}
\description{
For details on the methods see the documentation of the
corresponding classes.
}
\seealso{
\code{\link[=spflow_model-class]{spflow_model-class()}}
}
