% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R,
%   R/class_sp-network-meta.R
\name{variable_names}
\alias{variable_names}
\alias{variable_names<-}
\alias{variable_names,sp_network_nodes_pairs-method}
\title{Access the column names of the data inside \code{\link[=sp_network_classes]{sp_network_classes()}}}
\usage{
\S4method{variable_names}{sp_network_nodes_pairs}(object)

\S4method{variable_names}{sp_network_nodes_pairs}(object) <- value
}
\arguments{
\item{object}{A \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}} or \code{\link[=sp_network_pair-class]{sp_network_pair-class()}}}

\item{value}{A character of new variable names}
}
\description{
Access the column names of the data inside \code{\link[=sp_network_classes]{sp_network_classes()}}
}
\keyword{internal}
