% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_error.R
\name{custom_primitives}
\alias{custom_primitives}
\alias{is_one_of}
\alias{is_single_character}
\alias{is_single_logical}
\alias{is_single_numeric}
\alias{has_equal_elements}
\alias{has_distinct_elements}
\alias{none}
\title{A set of primitive functions to test conditions}
\usage{
is_one_of(.obj, .classes)

is_single_character(x)

is_single_logical(x)

is_single_numeric(x)

has_equal_elements(obj)

has_distinct_elements(obj)

none(x)
}
\value{
A logical
}
\description{
A set of primitive functions to test conditions
}
\author{
Lukas Dargel
}
\keyword{internal}
