% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{dat}
\alias{dat}
\alias{dat-set}
\alias{dat<-}
\title{Generic for accessing the data inside
\link[=sp_network_classes]{spflow network classes}}
\usage{
dat(object, ...)

dat(object, ...) <- value
}
\arguments{
\item{object}{A \link[=sp_network_classes]{spflow network classes}}

\item{value}{A data.frame to replace existing data}
}
\description{
For details on the methods see the documentation of the
corresponding classes.
}
\seealso{
\code{\link[=sp_network_classes]{sp_network_classes()}}
}
