% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-logdet.R
\name{derive_log_det_calculator}
\alias{derive_log_det_calculator}
\title{Fill the lookup table created by \code{\link[=multinom_lookup_template]{multinom_lookup_template()}} with values}
\usage{
derive_log_det_calculator(OW_traces = NULL, DW_traces = NULL, n_o, n_d, model)
}
\description{
Fill the lookup table created by \code{\link[=multinom_lookup_template]{multinom_lookup_template()}} with values
}
\keyword{internal}
