% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow_control.R
\name{enhance_spflow_control}
\alias{enhance_spflow_control}
\title{Enhance the estimation_control object for estimation}
\usage{
enhance_spflow_control(estimation_control, is_within)
}
\description{
Validate the input to flow control and add additional information related
to the matrix form expression of the model.
}
\keyword{internal}
