% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow_neighborhood.R
\name{expand_spflow_neighborhood}
\alias{expand_spflow_neighborhood}
\title{Derive the flow neighborhood matrices}
\usage{
expand_spflow_neighborhood(
  OW,
  DW,
  n_o = nrow(OW),
  n_d = nrow(DW),
  flow_indicator = NULL,
  model = "model_9"
)
}
\arguments{
\item{OW}{Origin neighborhood matrix}

\item{DW}{Destination neighborhood matrix}

\item{n_o}{A numeric indicating the number of origins}

\item{n_d}{A numeric indicating the number of destinations}

\item{flow_indicator}{A matrix of binary indicators for origin-destination
pairs that should be included in the model. When the argument is NULL or
when all entries of this matrix are different from zero we consider the
Cartesian product of all origins and destinations as OD pairs.}

\item{model}{A character indicating the model identifier}
}
\value{
A list of at most three (sparse) matrices with names
given by c("Wd", "Wo", "Ww").
}
\description{
Use the neighborhood matrices of origins and destinations to derive the
three neighborhood matrices of the origin-destination flows.
This function is used for simulations or for comparisons with the vectorized
formulations of the model.
}
\seealso{
Other spflow simulation functions: 
\code{\link{spatial_filter}()},
\code{\link{spflow_sim}()}
}
\concept{spflow simulation functions}
