% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{id}
\alias{id}
\alias{id-set}
\alias{id<-}
\title{Generic for accessing the ids of
\link[=spflow_network_classes]{spflow network classes}}
\usage{
id(object, ...)

id(object, ...) <- value
}
\arguments{
\item{object}{A \link[=spflow_network_classes]{spflow network class}}

\item{...}{Arguments to be passed to methods}

\item{value}{A character replacing the existing id}
}
\value{
A character, corresponding to the id
}
\description{
For details on the methods see the documentation of the
corresponding classes.

For details see the documentation of the corresponding classes.
}
\seealso{
\code{\link[=spflow_network_classes]{spflow_network_classes()}}
}
