% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{mcmc_results}
\alias{mcmc_results}
\title{Access the sampling results of a \code{\link[=spflow_model_mcmc-class]{spflow_model_mcmc-class()}}}
\usage{
mcmc_results(object)
}
\arguments{
\item{object}{A \code{\link[=spflow_model_mcmc-class]{spflow_model_mcmc-class()}}}
}
\value{
An mcmc object, containing the draws of the sampled parameters
}
\description{
For details see the documentation of the corresponding class.
}
\seealso{
\code{\link[=spflow_model_mcmc-class]{spflow_model_mcmc-class()}}
}
