% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-logdet.R
\name{multinom_lookup_template}
\alias{multinom_lookup_template}
\title{Create a lookup table for the multinomial coefficient}
\usage{
multinom_lookup_template(max_power, coef_names)
}
\description{
Create a lookup table for the multinomial coefficient
}
\keyword{internal}
