% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{neighborhood}
\alias{neighborhood}
\alias{neighborhood<-}
\title{Generic for accessing the neighborhood matrix inside
\link[=spflow_network_classes]{spflow network classes}}
\usage{
neighborhood(object, ...)

neighborhood(object) <- value
}
\arguments{
\item{object}{One of \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}}, \code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}

\item{...}{Arguments to be passed to methods}

\item{value}{A neighborhood matrix to replace the existing one}
}
\value{
A matrix (optionally sparse), representing the neighborhood links
betweem the nodes
}
\description{
For details on the methods see the documentation of the
corresponding classes.
}
\seealso{
\code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}} \code{\link[=sp_multi_network-class]{sp_multi_network-class()}}
}
