% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics_and_maybes.R
\name{nnodes}
\alias{nnodes}
\title{Generic for accessing the node count of
\link[=spflow_network_classes]{spflow network classes}}
\usage{
nnodes(object, ...)
}
\arguments{
\item{object}{One of \code{\link[=spflow_nodes-class]{spflow_nodes-class()}}, \code{\link[=spflow_pairs-class]{spflow_pairs-class()}}}

\item{...}{Arguments to be passed to methods}
}
\value{
A numeric, corresponding to the number of nodes
}
\description{
For details on the methods see the documentation of the
corresponding classes.
}
\seealso{
\code{\link[=spflow_nodes-class]{spflow_nodes-class()}} \code{\link[=spflow_pairs-class]{spflow_pairs-class()}}
}
