% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{nnodes}
\alias{nnodes}
\title{Generic for accessing the node count of
\link[=spflow_network_classes]{spflow network classes}}
\usage{
nnodes(object, ...)
}
\arguments{
\item{object}{One of \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}}, \code{\link[=sp_network_pair-class]{sp_network_pair-class()}}}

\item{...}{Arguments to be passed to methods}
}
\value{
A numeric, corresponding to the number of nodes
}
\description{
For details on the methods see the documentation of the
corresponding classes.
}
\seealso{
\code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}} \code{\link[=sp_network_pair-class]{sp_network_pair-class()}}
}
