% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{npairs}
\alias{npairs}
\title{Access the node pair count of a \code{\link[=sp_network_pair-class]{sp_network_pair-class()}}}
\arguments{
\item{object}{A \code{\link[=sp_network_pair-class]{sp_network_pair-class()}}}
}
\description{
For details see the documentation of the corresponding class.
}
