% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{npairs}
\alias{npairs}
\title{Generic for accessing the node pairs count of a
\code{\link[=sp_network_pair-class]{sp_network_pair-class()}}}
\usage{
npairs(object)
}
\arguments{
\item{object}{A \code{\link[=sp_network_pair-class]{sp_network_pair-class()}}}
}
\value{
A numeric, corresponding to the number of node-pairs
}
\description{
For details on the method see the documentation of the
corresponding class.
}
\seealso{
\code{\link[=sp_network_pair-class]{sp_network_pair-class()}}
}
