% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-data.R
\name{od_pair_information}
\alias{od_pair_information}
\title{Create an S4 object that contains information on origin-destination pairs}
\usage{
od_pair_information(origin_id, destination_id, pair_attributes = NULL)
}
\arguments{
\item{origin_id}{A character that serves as identifier for the origin network}

\item{destination_id}{A character that serves as identifier for the destination network}

\item{pair_attributes}{A list of matrices that contain information on node pairs (origins are in rows and destinations are in columns)}
}
\value{
An S4 class of type
}
\description{
Create an S4 object that contains information on origin-destination pairs
}
\seealso{
Other od_pair_information network_data: 
\code{\link{od_pair_information-class}}
}
\concept{od_pair_information network_data}
