% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{paris_data}
\alias{paris_data}
\alias{paris10km_nodes}
\alias{paris10km_node_pairs}
\alias{paris10km_mat_nb}
\title{Example data for commuting flows of paris}
\format{
An object of class \code{sf} (inherits from \code{data.table}, \code{data.frame}) with 71 rows and 6 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 5041 rows and 4 columns.

An object of class \code{list} of length 3.
}
\source{
The data combine different public sources:
\itemize{
\item \href{https://public.opendatasoft.com/explore/dataset/code-postal-code-insee-2015/table/?flg=fr}{Population, area, shape, ...}
\item \href{https://www.insee.fr/fr/statistiques/2021271}{Number of companies in each municipality}
\item \href{https://www.data.gouv.fr/fr/datasets/prix-moyen-au-m2-des-ventes-de-maisons-et-dappartements-par-commune-en-2019/}{Average price of housing}
\item \href{https://www.insee.fr/fr/statistiques/3560121}{Average Income}
}
}
\usage{
paris10km_nodes

paris10km_node_pairs

paris10km_mat_nb
}
\description{
The package includes an example data set that contains home-to-work
commuting flows for 71 municipalities around the centrer of Paris.
The data for the example is stored in three objects
}
\details{
\enumerate{
\item \code{paris10km_nodes} contains information on the municipalities.
It contains some socioeconomic variable, the identifier of the municipality
and a geometry column.
The geometry is a MULTIPOLYGON that describes the shape of the region.
\item \code{paris10km_node_pairs} contains information on pairs of municipalities.
It is a data.table with origin and destination identifiers and contains
the information on the size of the flow and the distance.
\item \code{paris10km_mat_nb} contains three sparse matrices that correspond to
alternative definitions of the neighborhood of the municipalities.
}
}
\keyword{datasets}
