% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{pull_member}
\alias{pull_member}
\title{Generic for accessing a \code{\link[=sp_network_pair-class]{sp_network_pair-class()}} or a
\code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}} inside a \code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}
\usage{
pull_member(object, ...)
}
\arguments{
\item{object}{A \code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}

\item{...}{Arguments to be passed to methods}
}
\value{
A \code{\link[=sp_network_pair-class]{sp_network_pair-class()}} or a \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}}
}
\description{
For details on the method see the documentation of the
corresponding class.
}
\seealso{
\code{\link[=sp_multi_network-class]{sp_multi_network-class()}}
}
