% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R,
%   R/class_sp-network-multi.R
\name{pull_neighborhood-deprecated}
\alias{pull_neighborhood-deprecated}
\alias{pull_neighborhood,sp_multi_network-method}
\title{Access the neighborhood matrix of the nodes inside a
\code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}
\usage{
pull_neighborhood(object, ...)

\S4method{pull_neighborhood}{sp_multi_network}(object, network_ids = NULL)
}
\arguments{
\item{object}{A \code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}

\item{network_ids}{A character vector indicating the ids of the
\code{\link[=sp_network_nodes]{sp_network_nodes()}} objects to extract from the
\code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}
}
\description{
Access the neighborhood matrix of the nodes inside a
\code{\link[=sp_multi_network-class]{sp_multi_network-class()}}
}
\seealso{
\code{\link{spflow-deprecated}}
}
\keyword{internal}
