% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R
\name{results<-}
\alias{results<-}
\title{Replacement method for internal purposes}
\usage{
results(object) <- value
}
\arguments{
\item{object}{A \code{\link[=spflow_model]{spflow_model()}}}

\item{value}{A data.frame of results}
}
\description{
Replacement method for internal purposes
}
\keyword{internal}
