% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R,
%   R/class_spflow-model-meta.R
\name{results_flat}
\alias{results_flat}
\alias{results_flat,spflow_model-method}
\title{Access results of a \code{\link[=spflow_model]{spflow_model()}} as a one row data.frame}
\usage{
results_flat(object, ...)

\S4method{results_flat}{spflow_model}(
  object,
  res_info = c("est", "sd"),
  cntrol_info = c("estimation_method")
)
}
\arguments{
\item{object}{spflow_model_meta}

\item{...}{Arguments to be passed to methods}
}
\description{
Access results of a \code{\link[=spflow_model]{spflow_model()}} as a one row data.frame

Reshaped version of the estimation results that can be used to compare
results in simulation studies.
}
\keyword{internal}
