% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_0_generics-and-maybe-classes.R,
%   R/class_sp-network-meta.R
\name{set_columns}
\alias{set_columns}
\alias{set_columns,sp_network_nodes_pairs-method}
\title{Add columns to the data in a \code{\link[=sp_network_nodes]{sp_network_nodes()}} or
\code{\link[=sp_network_pair]{sp_network_pair()}}}
\usage{
\S4method{set_columns}{sp_network_nodes_pairs}(object, new_cols)
}
\arguments{
\item{object}{One of; \code{\link[=sp_network_nodes]{sp_network_nodes()}}, \code{\link[=sp_network_pair]{sp_network_pair()}}}

\item{new_cols}{New columns to be added to existing data can be one of;
scalar, vector, data.frame}
}
\description{
Add columns to the data in a \code{\link[=sp_network_nodes]{sp_network_nodes()}} or
\code{\link[=sp_network_pair]{sp_network_pair()}}

Internal method that allows to overwrite columns in the data inside
\code{\link[=sp_network_classes]{sp_network_classes()}}
}
\details{
This function assigns by reference.
}
\keyword{internal}
