% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sp-network-multi.R
\docType{class}
\name{sp_multi_network-class}
\alias{sp_multi_network-class}
\alias{id,sp_multi_network-method}
\alias{dat,sp_multi_network-method}
\alias{pull_neighborhood,sp_multi_network-method}
\alias{pull_nodes,sp_multi_network-method}
\alias{pull_pairs,sp_multi_network-method}
\title{sp_multi_network Class}
\usage{
\S4method{id}{sp_multi_network}(object, what = cases)

\S4method{dat}{sp_multi_network}(object, network_id = NULL, network_pair_id = NULL)

\S4method{pull_neighborhood}{sp_multi_network}(object, network_ids = NULL)

\S4method{pull_nodes}{sp_multi_network}(object, network_ids = NULL)

\S4method{pull_pairs}{sp_multi_network}(object, network_pair_ids = NULL)
}
\arguments{
\item{what}{A character to indicating if ids of the single networks or the network
pairs should be retrieved; should be in \code{c("networks","network_pairs")}.}

\item{network_id}{A single character that correspond to the id of a
\code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}} inside the \code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}

\item{network_pair_id}{A single character that correspond to the id of a
\code{\link[=sp_network_pair-class]{sp_network_pair-class()}} inside the \code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}

\item{network_ids}{A character vector indicating the ids of the
\code{\link[=sp_network_nodes]{sp_network_nodes()}} objects to extract from the
\code{\link[=sp_multi_network-class]{sp_multi_network-class()}}}

\item{network_pair_ids}{A character vector of ids for contained \code{\link[=sp_network_pair]{sp_network_pair()}} objects}
}
\description{
A S4 class that gathers information on one or multiple networks
\code{\link[=sp_network_nodes]{sp_network_nodes()}} and origin-destination pairs \code{\link[=sp_network_pair]{sp_network_pair()}}.
The class is constructed with the \code{\link[=sp_multi_network]{sp_multi_network()}} function.
}
\section{Slots}{

\describe{
\item{\code{networks}}{A list of \code{\link[=sp_network_nodes-class]{sp_network_nodes-class()}} objects}

\item{\code{network_pairs}}{A list of \code{\link[=sp_network_pair-class]{sp_network_pair-class()}} objects}

\item{\code{completeness}}{A data.frame that provides summary information on the matching between
the nodes and node pairs that are provided to the sp_multi_network class}
}}

\examples{
## access the id of a networks or network_pairs inside a multi network

id(multi_net_usa_ge,"networks")
id(multi_net_usa_ge,"network_pairs")

## access the data of a network or a network_pair inside a multi_network

dat(multi_net_usa_ge,network_id = "ge") # extract data of nodes
dat(multi_net_usa_ge,network_pair_id = "ge_ge") # extract data of pairs

## access a network inside a multi_network

pull_nodes(multi_net_usa_ge,"ge")
pull_nodes(multi_net_usa_ge,"usa")

## access a network inside a multi_network

pull_pairs(multi_net_usa_ge,"ge_ge")
pull_pairs(multi_net_usa_ge,"usa_usa")

}
\seealso{
Other spflow network objects: 
\code{\link{sp_multi_network}},
\code{\link{sp_network_classes}},
\code{\link{sp_network_nodes-class}},
\code{\link{sp_network_nodes}},
\code{\link{sp_network_pair-class}},
\code{\link{sp_network_pair}}
}
\concept{spflow network objects}
\keyword{internal}
