% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_network-data.R
\name{sp_multi_network}
\alias{sp_multi_network}
\title{Create an S4 class that contains \link{sp_network} and \link{sp_network_pair} for one or multiple networks}
\usage{
sp_multi_network(...)
}
\arguments{
\item{...}{objects of of type \link{sp_network} and \link{sp_network_pair}}
}
\value{
A S4 network data object
}
\description{
Create an S4 class that contains \link{sp_network} and \link{sp_network_pair} for one or multiple networks
}
