% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sp-network-multi.R
\name{sp_multi_network}
\alias{sp_multi_network}
\title{Create an S4 class that contains \code{\link[=sp_network_nodes]{sp_network_nodes()}} and \code{\link[=sp_network_pair]{sp_network_pair()}} for one or multiple networks}
\usage{
sp_multi_network(...)
}
\arguments{
\item{...}{objects of type \code{\link[=sp_network_nodes]{sp_network_nodes()}} and \code{\link[=sp_network_pair]{sp_network_pair()}}}
}
\value{
An S4 class of type \code{\link[=sp_multi_network-class]{sp_multi_network-class()}}
}
\description{
Create an S4 class that contains \code{\link[=sp_network_nodes]{sp_network_nodes()}} and \code{\link[=sp_network_pair]{sp_network_pair()}} for one or multiple networks
}
\examples{
sp_multi_network() # empty
sp_multi_network(germany_net,usa_net) # two networks, no pairs
}
\seealso{
Other Constructors for spflow network classes: 
\code{\link{sp_network_nodes}},
\code{\link{sp_network_pair}},
\code{\link{spflow_network_classes}}
}
\concept{Constructors for spflow network classes}
