% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_od-pair-information.R
\name{sp_network_pair}
\alias{sp_network_pair}
\title{Create an S4 object that contains information on origin-destination pairs}
\usage{
sp_network_pair(origin_id, destination_id, pair_data = NULL)
}
\arguments{
\item{origin_id}{A character that serves as identifier for the origin network}

\item{destination_id}{A character that serves as identifier for the destination network}

\item{pair_data}{A list of matrices that contain information on node pairs (origins are in rows and destinations are in columns)}
}
\value{
An S4 class of type
}
\description{
Create an S4 object that contains information on origin-destination pairs
}
\concept{sp_network_pair sp_multi_network}
