% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_sp-network-pair.R
\name{sp_network_pair}
\alias{sp_network_pair}
\title{Create an S4 object that contains information on origin-destination pairs}
\usage{
sp_network_pair(
  origin_network_id,
  destination_network_id,
  node_pair_data = NULL,
  origin_key_column = NULL,
  origin_node_count = NULL,
  destination_key_column = NULL,
  destination_node_count = NULL
)
}
\arguments{
\item{origin_network_id}{A character that serves as identifier for the origin network}

\item{destination_network_id}{A character that serves as identifier for the destination network}

\item{node_pair_data}{}
}
\value{
An S4 class of type
}
\description{
Create an S4 object that contains information on origin-destination pairs
}
\concept{sp_network_pair sp_multi_network}
