% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-package.R
\docType{package}
\name{spflow-package}
\alias{spflow}
\alias{spflow-package}
\title{spflow: A package for spatial econometric interaction models.}
\description{
The foo package provides function to estimate origin destination flows with
that exhibit spatial autocorrelation.
}
\section{Network data}{

The package provides a new class for representing network data.
}

\section{Model estimation}{

The package provides functions to estimate spatial interaction models.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukeCe/spflow}
  \item Report bugs at \url{https://github.com/LukeCe/spflow/issues}
}

}
\author{
\strong{Maintainer}: Lukas Dargel \email{lukas.dargel@mailbox.org} (\href{https://orcid.org/0000-0002-4839-506X}{ORCID})

}
\keyword{internal}
