% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-control.R
\name{spflow_control}
\alias{spflow_control}
\title{Create a control object for the estimation with \code{\link[=spflow]{spflow()}}}
\usage{
spflow_control(
  estimation_method = "s2sls",
  hessian_method = "mixed",
  sdm_variables = "same",
  instrumental_variables = "same"
)
}
\arguments{
\item{estimation_method}{A character which indicates the estimation method}

\item{hessian_method}{A character which indicates the method for hessian calculation}

\item{sdm_variables}{A formula which can be used explicitly delcare the variables that should be used as sdm variables.}

\item{instrumental_variables}{A formula which can be used explicitly delcare the variables that should be used as instruments during s2sls estimation.}
}
\value{
A list of control parameters
}
\description{
This function creates a list to fine tune the estimation of a spatial
interaction model with \code{\link[=spflow]{spflow()}}.
}
\seealso{
spflow
}
