% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generics_and_maybes.R,
%   R/class_spflow_network_multi.R, R/class_spflow_model.R
\name{spflow_map}
\alias{spflow_map}
\alias{spflow_map,spflow_network_multi-method}
\alias{spflow_map,spflow_model-method}
\title{Geographic representation of flows}
\usage{
spflow_map(object, ...)

\S4method{spflow_map}{spflow_network_multi}(object, id_spflow_pairs = id(object)[["pairs"]][[1]], ..., flow_var)

\S4method{spflow_map}{spflow_model}(object, ..., flow_type = "resid", add_title = TRUE)
}
\arguments{
\item{object}{A \code{\link[=spflow_network_multi-class]{spflow_network_multi-class()}} or a \code{\link[=spflow_model-class]{spflow_model-class()}}}

\item{...}{arguments passed to methods and \code{map_flows()}}

\item{id_spflow_pairs}{A character indicating the id of a \code{\link[=spflow_network_pairs]{spflow_network_pairs()}} (only relevant if
the \code{\link[=spflow_network_multi]{spflow_network_multi()}} contains multiple \code{spflow_network_pairs}-objects:
defaults to the of them)}

\item{flow_var}{A character, indicating one variable from the network pair data}

\item{flow_type}{A character indicating what to values to show.
Should be one of c("resid", "fitted", "actual").}

\item{add_title}{A logical, if \code{TRUE} the flow_type is added as title.}
}
\description{
This generic is used as an interface to \code{map_flows()}, where all information
available in the object is extracted and provided as arguments.
}
\examples{

 # Used with a spflow_network_multi ...
 spflow_map(multi_net_usa_ge, "ge_ge",flow_var = "y9")


 # Used with a model...
 gravity_ge <- spflow(
   y1 ~ . + P_(DISTANCE),
   multi_net_usa_ge,
   "ge_ge",
   spflow_control(model = "model_1"))
 spflow_map(gravity_ge)
 spflow_map(gravity_ge, flow_type = "fitted")
 spflow_map(gravity_ge, flow_type = "actual")

}
\author{
Lukas Dargel
}
