% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spflow-model-matrix_1.R
\name{spflow_model_matrix}
\alias{spflow_model_matrix}
\title{Generate design matrices for the spatial interaction model}
\usage{
spflow_model_matrix(
  sp_multi_network,
  network_pair_id,
  formula_parts,
  estim_control
)
}
\arguments{
\item{sp_multi_network}{A \code{\link[=sp_multi_network]{sp_multi_network()}} object that contains information on the origins,
and the destinations and their neighborhood structure}

\item{network_pair_id}{A character indicating the id of a \code{\link[=sp_network_pair]{sp_network_pair()}} (only relevant if
the \code{\link[=sp_multi_network]{sp_multi_network()}} contains multiple \code{sp_network_pair}-objects:
defaults to the of them)}
}
\value{
A list of design matrices for spatial interaction model
}
\description{
Creates model matrices from a \code{\link[=sp_multi_network]{sp_multi_network()}} network.
For efficiency the relational representation of origin and destination data
is preserved.
}
\keyword{internal}
