% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_multinet.R
\docType{class}
\name{spflow_multinet-class}
\alias{spflow_multinet-class}
\alias{id,spflow_multinet-method}
\alias{complete_pairs,spflow_multinet-method}
\alias{dat,spflow_multinet-method}
\alias{dat<-,spflow_multinet-method}
\alias{neighborhood,spflow_multinet-method}
\alias{nnodes,spflow_multinet-method}
\alias{npairs,spflow_multinet-method}
\alias{pull_member,spflow_multinet-method}
\title{Class spflow_multinet}
\usage{
\S4method{id}{spflow_multinet}(object)

\S4method{complete_pairs}{spflow_multinet}(object, network_pair_ids = id(object)[["network_pairs"]])

\S4method{dat}{spflow_multinet}(object, .id)

\S4method{dat}{spflow_multinet}(object, .id) <- value

\S4method{neighborhood}{spflow_multinet}(object, .id = 1)

\S4method{nnodes}{spflow_multinet}(object, .id)

\S4method{npairs}{spflow_multinet}(object, .id)

\S4method{pull_member}{spflow_multinet}(object, .id = NULL)
}
\arguments{
\item{object}{spflow_multinet-class}

\item{network_pair_ids}{A character indicating of one or several \code{\link[=spflow_pairs]{spflow_pairs()}} objects}

\item{.id}{A character, corresponding to the id of a \code{spflow_nodes()} object
contained in the \code{spflow_multinet}}

\item{value}{A data.frame to replace the existing data}
}
\description{
A S4 class that gathers information on one or multiple networks
\code{\link[=spflow_nodes]{spflow_nodes()}} and origin-destination pairs \code{\link[=spflow_pairs]{spflow_pairs()}}.
The class is constructed with the \code{\link[=spflow_multinet]{spflow_multinet()}} function.
}
\section{Slots}{

\describe{
\item{\code{networks}}{A list of \code{\link[=spflow_nodes-class]{spflow_nodes-class()}} objects}

\item{\code{network_pairs}}{A list of \code{\link[=spflow_pairs-class]{spflow_pairs-class()}} objects}
}}

\examples{
## access the id of a networks or network_pairs inside a multi network

id(multi_net_usa_ge)$networks
id(multi_net_usa_ge)$network_pairs

## access the data of a network or a network_pair inside a multi_network

dat(multi_net_usa_ge, "ge")    # extract data of nodes
dat(multi_net_usa_ge, "ge_ge") # extract data of pairs

## access the data of a network or a network_pair inside a multi_network

dat(multi_net_usa_ge, "ge")    # extract data of nodes
dat(multi_net_usa_ge, "ge_ge") # extract data of pairs

## access spflow_nodes or spflow_pairs inside a spflow_multinet

pull_member(multi_net_usa_ge, .id = "ge")
pull_member(multi_net_usa_ge, .id = "usa")
pull_member(multi_net_usa_ge, .id ="ge_ge")

}
\seealso{
Other spflow network classes: 
\code{\link{spflow_network_classes}},
\code{\link{spflow_nodes-class}},
\code{\link{spflow_pairs-class}}
}
\concept{spflow network classes}
