% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_spflow_network_multi.R
\name{spflow_network_multi}
\alias{spflow_network_multi}
\title{Constructor for the \code{\link[=spflow_network_multi-class]{spflow_network_multi-class()}}}
\usage{
spflow_network_multi(...)
}
\arguments{
\item{...}{objects of type \code{\link[=spflow_network]{spflow_network()}} and \code{\link[=spflow_network_pair]{spflow_network_pair()}}}
}
\value{
An S4 class of type \code{\link[=spflow_network_multi-class]{spflow_network_multi-class()}}
}
\description{
Constructor for the \code{\link[=spflow_network_multi-class]{spflow_network_multi-class()}}
}
\examples{
spflow_network_multi() # empty
spflow_network_multi(germany_net,usa_net) # two nodes, no pairs
}
\seealso{
spflow_network_classes
}
